refactor : Add data.js and modify components accordingly
