+++ src/components/JustDropped.js
@@ -1,6 +1,6 @@
 import { useState } from "react";
-import "../css/JustDropped.css"
-import justDroppedData from "../justDroppedData.js"
+import "../css/JustDropped.css";
+import { justDroppedData } from "../constants/data.js";
 import SaleItem from "./SaleItem.js";
 import ShowMoreBtn from "./ShowMoreBtn.js";
