+++ src/components/MostPopular.js
@@ -1,9 +1,8 @@
 import { useState } from "react";
 import EventBanner from './EventBanner.js';
-import eventBannerData from '../eventBannerData.js';
-import mostPopularData from "../mostPopularData.js";
+import { eventBannerData, mostPopularData } from "../constants/data.js";
 import SaleItem from "./SaleItem.js";
-import "../css/MostPopular.css"
+import "../css/MostPopular.css";
 import ShowMoreBtn from "./ShowMoreBtn.js";
 
 function MostPopular() {
