+++ src/constants/data.js
@@ -1,0 +1,2542 @@
+export const eventData = [
+  {
+    id: 0,
+    src: "/images/eventBanner/bg_0.webp",
+    bgColor: "#A9A099",
+    activeColor: true
+  },
+  {
+    id: 1,
+    src: "/images/eventBanner/bg_1.webp",
+    bgColor: "#9A9196",
+    activeColor: true
+  },
+  {
+    id: 2,
+    src: "/images/eventBanner/bg_2.webp",
+    bgColor: "#868686",
+    activeColor: true
+  },
+  {
+    id: 3,
+    src: "/images/eventBanner/bg_3.webp",
+    bgColor: "#9796CE",
+    activeColor: true
+  }
+  ,{
+    id: 4,
+    src: "/images/eventBanner/bg_4.webp",
+    bgColor: "#828177",
+    activeColor: true
+  },
+  {
+    id: 5,
+    src: "/images/eventBanner/bg_5.webp",
+    bgColor: "#BABABA",
+    activeColor: true
+  },
+  {
+    id: 6,
+    src: "/images/eventBanner/bg_6.webp",
+    bgColor: "#E5E5E5",
+    activeColor: false
+  },
+  {
+    id: 7,
+    src: "/images/eventBanner/bg_7.webp",
+    bgColor: "#7B7678",
+    activeColor: true
+  },
+  {
+    id: 8,
+    src: "/images/eventBanner/bg_8.webp",
+    bgColor: "#E0FA9D",
+    activeColor: false
+  }
+  ,{
+    id: 9,
+    src: "/images/eventBanner/bg_9.webp",
+    bgColor: "#e5e5e5",
+    activeColor: false
+  },
+  {
+    id: 10,
+    src: "/images/eventBanner/bg_10.webp",
+    bgColor: "#D5DEED",
+    activeColor: false
+  },
+  {
+    id: 11,
+    src: "/images/eventBanner/bg_11.webp",
+    bgColor: "#86807E",
+    activeColor: true
+  },
+  {
+    id: 12,
+    src: "/images/eventBanner/bg_12.webp",
+    bgColor: "#025396",
+    activeColor: true
+  },
+  {
+    id: 13,
+    src: "/images/eventBanner/bg_13.webp",
+    bgColor: "#EFEFEF",
+    activeColor: false
+  }
+  ,{
+    id: 14,
+    src: "/images/eventBanner/bg_14.webp",
+    bgColor: "#030802",
+    activeColor: true
+  },
+  {
+    id: 15,
+    src: "/images/eventBanner/bg_15.webp",
+    bgColor: "#BFAD6E",
+    activeColor: true
+  },
+  {
+    id: 16,
+    src: "/images/eventBanner/bg_16.webp",
+    bgColor: "#DDDBC8",
+    activeColor: true
+  },
+  {
+    id: 17,
+    src: "/images/eventBanner/bg_17.webp",
+    bgColor: "#251313",
+    activeColor: true
+  },
+  {
+    id: 18,
+    src: "/images/eventBanner/bg_18.webp",
+    bgColor: "#0A0A07",
+    activeColor: true
+  }
+  ,{
+    id: 19,
+    src: "/images/eventBanner/bg_19.webp",
+    bgColor: "#0D0D0E",
+    activeColor: true
+  },
+  {
+    id: 20,
+    src: "/images/eventBanner/bg_20.webp",
+    bgColor: "#ACA79C",
+    activeColor: true
+  },
+  {
+    id: 21,
+    src: "/images/eventBanner/bg_21.webp",
+    bgColor: "#198dfa",
+    activeColor: true
+  },
+  {
+    id: 22,
+    src: "/images/eventBanner/bg_22.webp",
+    bgColor: "#252829",
+    activeColor: true
+  },
+  {
+    id: 23,
+    src: "/images/eventBanner/bg_23.webp",
+    bgColor: "#777E82",
+    activeColor: true
+  }
+  ,{
+    id: 24,
+    src: "/images/eventBanner/bg_24.webp",
+    bgColor: "#F9E5EE",
+    activeColor: false
+  },
+  {
+    id: 25,
+    src: "/images/eventBanner/bg_25.webp",
+    bgColor: "#50575A",
+    activeColor: true
+  },
+  {
+    id: 26,
+    src: "/images/eventBanner/bg_26.webp",
+    bgColor: "#C7AE9B",
+    activeColor: true
+  },
+  {
+    id: 27,
+    src: "/images/eventBanner/bg_27.webp",
+    bgColor: "#f5f5f5",
+    activeColor: false
+  },
+  {
+    id: 28,
+    src: "/images/eventBanner/bg_28.webp",
+    bgColor: "#E0E0E0",
+    activeColor: false
+  }
+  ,{
+    id: 29,
+    src: "/images/eventBanner/bg_29.webp",
+    bgColor: "#C8CDC8",
+    activeColor: true
+  },
+  {
+    id: 30,
+    src: "/images/eventBanner/bg_30.webp",
+    bgColor: "#D4E9FA",
+    activeColor: false
+  },
+  {
+    id: 31,
+    src: "/images/eventBanner/bg_31.webp",
+    bgColor: "#57251A",
+    activeColor: true
+  },
+  {
+    id: 32,
+    src: "/images/eventBanner/bg_32.webp",
+    bgColor: "#2BB165",
+    activeColor: true
+  },
+  {
+    id: 33,
+    src: "/images/eventBanner/bg_33.webp",
+    bgColor: "#000000",
+    activeColor: true
+  }
+  ,{
+    id: 34,
+    src: "/images/eventBanner/bg_34.webp",
+    bgColor: "#AA9A7C",
+    activeColor: true
+  },
+  {
+    id: 35,
+    src: "/images/eventBanner/bg_35.webp",
+    bgColor: "#F5F6F3",
+    activeColor: false
+  },
+  {
+    id: 36,
+    src: "/images/eventBanner/bg_36.webp",
+    bgColor: "#C387AF",
+    activeColor: true
+  },
+  {
+    id: 37,
+    src: "/images/eventBanner/bg_37.webp",
+    bgColor: "#020202",
+    activeColor: true
+  },
+  {
+    id: 38,
+    src: "/images/eventBanner/bg_38.webp",
+    bgColor: "#A9A69D",
+    activeColor: true
+  },
+  {
+    id: 39,
+    src: "/images/eventBanner/bg_39.webp",
+    bgColor: "#d8dee2",
+    activeColor: false
+  },
+  {
+    id: 40,
+    src: "/images/eventBanner/bg_40.webp",
+    bgColor: "#222222",
+    activeColor: true
+  },
+  {
+    id: 41,
+    src: "/images/eventBanner/bg_41.webp",
+    bgColor: "#d8dee2",
+    activeColor: false
+  },
+]
+
+export const eventDetailData = [
+  {
+    id: 0,
+    type: "webp",
+    src: "/images/eventDetailBtn/bg_0.gif",
+    content: "내 폰 시세"
+  },
+  {
+    id: 1,
+    type: "jpeg",
+    src: "/images/eventDetailBtn/bg_1.jpeg",
+    content: "크림 드로우"
+  },
+  {
+    id: 2,
+    type: "jpg",
+    src: "/images/eventDetailBtn/bg_2.jpg",
+    content: "남성 추천"
+  },
+  {
+    id: 3,
+    type: "jpg",
+    src: "/images/eventDetailBtn/bg_3.jpg",
+    content: "여성 추천"
+  },
+  {
+    id: 4,
+    type: "jpg",
+    src: "/images/eventDetailBtn/bg_4.jpg",
+    content: "색다른 추천"
+  },
+  {
+    id: 5,
+    type: "jpg",
+    src: "/images/eventDetailBtn/bg_5.jpg",
+    content: "정가 아래"
+  },
+  {
+    id: 6,
+    type: "jpg",
+    src: "/images/eventDetailBtn/bg_6.jpg",
+    content: "시즌 오프 ~70%"
+  },
+  {
+    id: 7,
+    type: "jpg",
+    src: "/images/eventDetailBtn/bg_7.jpg",
+    content: "역시즌 최대 82%"
+  },
+  {
+    id: 8,
+    type: "jpg",
+    src: "/images/eventDetailBtn/bg_8.jpg",
+    content: "2월 혜택"
+  },
+  {
+    id: 9,
+    type: "jpg",
+    src: "/images/eventDetailBtn/bg_9.jpg",
+    content: "눕시 & 아식스 선물"
+  }
+]
+
+export const justDroppedData= [
+  {
+    id: 0,
+    brand: "Vandy The Pink",
+    name: "Vandy The Pink x Alpha Industries Bomber Jacket Lead Gray",
+    price: "489,000",
+    saleCount: 34,
+    saleType: true,
+    src: "/images/justDropped/img0.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 1,
+    brand: "New Balance",
+    name: "New Balance 1906R Lunar New Year Pink",
+    price: "489,000",
+    saleCount: 12,
+    saleType: true,
+    src: "/images/justDropped/img1.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 2,
+    brand: "Palace",
+    name: "Palace Korea Varsity Jacket Black - 24SS",
+    price: "1,080,000",
+    saleCount: 19,
+    saleType: true,
+    src: "/images/justDropped/img2.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 3,
+    brand: "Human Made",
+    name: "Human Made Reversible Yokosuka Jacket Black",
+    price: "1,600,000",
+    saleCount: 1,
+    saleType: true,
+    src: "/images/justDropped/img3.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 4,
+    brand: "Blitzway",
+    name: "[예약 배송] Blitzway Disney D100 Titanium Mickey Mouse",
+    price: "2,899,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/justDropped/img4.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 5,
+    brand: "Asics",
+    name: "Asics Gel-Quantum Kinetic Mid Grey Pure Silver",
+    price: "332,000",
+    saleCount: 27,
+    saleType: true,
+    src: "/images/justDropped/img5.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 6,
+    brand: "Converse",
+    name: "Converse x Feng Chen Wang Chuck 70 Ox 2-in-1 Brown Rice",
+    price: "228,000",
+    saleCount: 63,
+    saleType: true,
+    src: "/images/justDropped/img6.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 7,
+    brand: "Nike",
+    name: "Nike Dunk Low SP Brazil 2024",
+    price: "160,000",
+    saleCount: 31,
+    saleType: true,
+    src: "/images/justDropped/img7.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 8,
+    brand: "Hoka",
+    name: "Hoka Bondi 8 Shifting Sand Vanilla",
+    price: "219,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/justDropped/img8.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 9,
+    brand: "Starbucks",
+    name: "Starbucks Be Mine Bear Fabric Keychain",
+    price: "36,000",
+    saleCount: 113,
+    saleType: true,
+    src: "/images/justDropped/img9.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 10,
+    brand: "New Balance",
+    name: "New Balance 860v2 Optic White",
+    price: "162,000",
+    saleCount: 63,
+    saleType: true,
+    src: "/images/justDropped/img10.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 11,
+    brand: "Adidas",
+    name: "(W) Adidas Samba OG Core Black Wonder White",
+    price: "145,000",
+    saleCount: 201,
+    saleType: true,
+    src: "/images/justDropped/img11.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 12,
+    brand: "Jordan",
+    name: "Jordan 1 Retro High OG Black White",
+    price: "199,000",
+    saleCount: 6,
+    saleType: true,
+    src: "/images/justDropped/img12.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 13,
+    brand: "Nike",
+    name: "Nike x Off-White Allover Print Jersey Kelly Green - Asia",
+    price: "200,000",
+    saleCount: 156,
+    saleType: true,
+    src: "/images/justDropped/img13.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 14,
+    brand: "Stussy",
+    name: "Stussy Wave Dye Beach Shell Jacket Black",
+    price: "323,000",
+    saleCount: 158,
+    saleType: true,
+    src: "/images/justDropped/img14.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+]
+
+export const topBrandDetailData = [
+  {
+    id: 0,
+    type: "jpg",
+    src: "/images/brands/Nike.jpg",
+    content: "나이키"
+  },
+  {
+    id: 1,
+    type: "jpg",
+    src: "/images/brands/Stussy.jpg",
+    content: "스투시"
+  },
+  {
+    id: 2,
+    type: "jpg",
+    src: "/images/brands/Supreme.jpg",
+    content: "슈프림"
+  },
+  {
+    id: 3,
+    type: "jpg",
+    src: "/images/brands/Jordan.jpg",
+    content: "조던"
+  },
+  {
+    id: 4,
+    type: "jpg",
+    src: "/images/brands/NewBalance.jpg",
+    content: "뉴발란스"
+  },
+  {
+    id: 5,
+    type: "jpg",
+    src: "/images/brands/ArcTeryx.jpg",
+    content: "아크테릭스"
+  },
+  {
+    id: 6,
+    type: "jpg",
+    src: "/images/brands/Asics.jpg",
+    content: "아식스"
+  },
+  {
+    id: 7,
+    type: "jpg",
+    src: "/images/brands/CarharttWIP.jpg",
+    content: "칼하트 WIP"
+  },
+  {
+    id: 8,
+    type: "jpg",
+    src: "/images/brands/HumanMade.jpg",
+    content: "휴먼 메이드"
+  },
+  {
+    id: 9,
+    type: "jpg",
+    src: "/images/brands/Adidas.jpg",
+    content: "아디다스"
+  },
+  {
+    id: 10,
+    type: "jpg",
+    src: "/images/brands/IABStudio.jpg",
+    content: "아이앱 스튜디오"
+  },
+  {
+    id: 11,
+    type: "jpg",
+    src: "/images/brands/AderError.jpg",
+    content: "아더에러"
+  },
+  {
+    id: 12,
+    type: "jpg",
+    src: "/images/brands/Palace.jpg",
+    content: "팔라스"
+  },
+  {
+    id: 13,
+    type: "jpg",
+    src: "/images/brands/StoneIsland.jpg",
+    content: "스톤 아일랜드"
+  },
+  {
+    id: 14,
+    type: "jpg",
+    src: "/images/brands/PoloRalphLauren.jpg",
+    content: "폴로 랄프 로렌"
+  }
+]
+
+export const eventBannerData = [
+  {
+    id: 0,
+    src: "/images/eventBanner/banner_0.webp",
+    bgColor: "#6934FF"
+  },
+  {
+    id: 1,
+    src: "/images/eventBanner/banner_1.webp",
+    bgColor: "#6B87A7"
+  },
+  {
+    id: 2,
+    src: "/images/eventBanner/banner_2.webp",
+    bgColor: "#4567E3"
+  },
+  {
+    id: 3,
+    src: "/images/eventBanner/banner_3.webp",
+    bgColor: "#F9E95F"
+  },
+  {
+    id: 4,
+    src: "/images/eventBanner/banner_4.webp",
+    bgColor: "rgb(158, 163, 149)"
+  },
+  {
+    id: 5,
+    src: "/images/eventBanner/banner_5.webp",
+    bgColor: "rgb(110, 127, 119)"
+  },
+  {
+    id: 6,
+    src: "/images/eventBanner/banner_6.webp",
+    bgColor: "rgb(110, 111, 115)"
+  },
+  {
+    id: 7,
+    src: "/images/eventBanner/banner_7.webp",
+    bgColor: "rgb(93, 137, 147)"
+  },
+  {
+    id: 8,
+    src: "/images/eventBanner/banner_8.webp",
+    bgColor: "rgb(138, 131, 131)"
+  },
+]
+
+export const mostPopularData= [
+  {
+    id: 0,
+    brand: "Nike",
+    name: "Nike Air Force 1 '07 WB Flax",
+    price: "144,000",
+    saleCount: 147000,
+    saleType: true,
+    src: "/images/mostPopular/img0.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 1,
+    brand: "Jordan",
+    name: "Jordan 4 Retro Bred Reimagined",
+    price: "287,000",
+    saleCount: 236,
+    saleType: true,
+    src: "/images/mostPopular/img1.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 2,
+    brand: "Adidas",
+    name: "Adidas Samba OG Cloud White",
+    price: "117,000",
+    saleCount: 58000,
+    saleType: true,
+    src: "/images/mostPopular/img2.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 3,
+    brand: "Nike",
+    name: "Nike Air Force 1 '07 Low White",
+    price: "123,000",
+    saleCount: 438000,
+    saleType: true,
+    src: "/images/mostPopular/img3.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 4,
+    brand: "Apple",
+    name: "Apple Airpods Pro 2nd Gen with MagSafe Charging Case USB C Type (Korean Ver.)",
+    price: "306,000",
+    saleCount: 6655,
+    saleType: true,
+    src: "/images/mostPopular/img4.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 5,
+    brand: "Adidas",
+    name: "(W) Adidas Spezial Handball Clear Pink",
+    price: "129,000",
+    saleCount: 6777,
+    saleType: true,
+    src: "/images/mostPopular/img5.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 6,
+    brand: "Adidas",
+    name: "Adidas Samba OG Black White Gum",
+    price: "119,000",
+    saleCount: 52000,
+    saleType: true,
+    src: "/images/mostPopular/img6.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 7,
+    brand: "New Balance",
+    name: "New Balance 993 Made in USA Grey - D Standard",
+    price: "310,000",
+    saleCount: 64000,
+    saleType: true,
+    src: "/images/mostPopular/img7.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 8,
+    brand: "Nike",
+    name: "(W) Nike Zoom Vomero 5 Photon Dust and Metallic Silver",
+    price: "172,000",
+    saleCount: 15000,
+    saleType: true,
+    src: "/images/mostPopular/img8.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 9,
+    brand: "Adidas",
+    name: "(W) Adidas Spezial Handball Core Black Clear Pink",
+    price: "147,000",
+    saleCount: 1406,
+    saleType: true,
+    src: "/images/mostPopular/img9.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+
+  {
+    id: 10,
+    brand: "Mihara Yasuhiro",
+    name: "Maison Mihara Yasuhiro Peterson OG Sole Canvas Low Sneakers Black",
+    price: "250,000",
+    saleCount: 24000,
+    saleType: true,
+    src: "/images/mostPopular/img10.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 11,
+    brand: "Asics",
+    name: "Asics Gel-Kayano 14 Cream Black",
+    price: "224,000",
+    saleCount: 201,
+    saleType: true,
+    src: "/images/mostPopular/img11.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 12,
+    brand: "Nike",
+    name: "Nike Air Force 1 '07 Low Triple Black",
+    price: "124,000",
+    saleCount: 71000,
+    saleType: true,
+    src: "/images/mostPopular/img12.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 13,
+    brand: "Stussy",
+    name: "Stussy Basic Stussy Hoodie Black 2023",
+    price: "222,000",
+    saleCount: 9520,
+    saleType: true,
+    src: "/images/mostPopular/img13.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 14,
+    brand: "Nike",
+    name: "Nike x Ambush Air Force 1 Low SP Black",
+    price: "298,000",
+    saleCount: 7331,
+    saleType: true,
+    src: "/images/mostPopular/img14.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 15,
+    brand: "Nike",
+    name: "(W) Nike Air Force 1 '07 Low White",
+    price: "110,000",
+    saleCount: 151000,
+    saleType: true,
+    src: "/images/mostPopular/img15.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 16,
+    brand: "New Balance",
+    name: "New Balance 2002R Grey",
+    price: "159,000",
+    saleCount: 66000,
+    saleType: true,
+    src: "/images/mostPopular/img16.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 17,
+    brand: "Nike",
+    name: "Nike Zoom Vomero 5 PRM Light Iron Ore and Flat Fewter",
+    price: "150,000",
+    saleCount: 6,
+    saleType: true,
+    src: "/images/mostPopular/img17.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 18,
+    brand: "Nike",
+    name: "(W) Nike Air Force 1 '07 Cacao Wow",
+    price: "234,000",
+    saleCount: 4018,
+    saleType: true,
+    src: "/images/mostPopular/img18.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 19,
+    brand: "Nike",
+    name: "(W) Nike Zoom Vomero 5 Cobblestone and Flat Pewter",
+    price: "159,000",
+    saleCount: 34000,
+    saleType: true,
+    src: "/images/mostPopular/img19.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+]
+
+export const coordinationData = [
+  {
+    id: 0,
+    thumbnailSrc: "/images/coordiImg/user_yeon2.jpeg",
+    thumbnailType: "jpeg",
+    userId: "y.e.o.n2",
+    coordiImgSrc: "/images/coordiImg/userCoordi0.jpegtype=m_webp",
+    coordiImgType: "jpegtype=m_webp"
+  },
+  {
+    id: 1,
+    thumbnailSrc: "/images/coordiImg/user_sonyunko.jpeg",
+    thumbnailType: "jpeg",
+    userId: "sonyunko",
+    coordiImgSrc: "/images/coordiImg/userCoordi1.jpegtype=m_webp",
+    coordiImgType: "jpegtype=m_webp"
+  },
+  {
+    id: 2,
+    thumbnailSrc: "/images/coordiImg/user_gogoomee.jpeg",
+    thumbnailType: "jpeg",
+    userId: "gogoomee",
+    coordiImgSrc: "/images/coordiImg/userCoordi2.jpegtype=m_webp",
+    coordiImgType: "jpegtype=m_webp"
+  },
+  {
+    id: 3,
+    thumbnailSrc: "/images/coordiImg/user_everywear.jpeg",
+    thumbnailType: "jpeg",
+    userId: "everywear",
+    coordiImgSrc: "/images/coordiImg/userCoordi3.jpegtype=m_webp",
+    coordiImgType: "jpegtype=m_webp"
+  },
+  {
+    id: 4,
+    thumbnailSrc: "/images/coordiImg/user_heeeseok.jpeg",
+    thumbnailType: "jpeg",
+    userId: "heee.seok",
+    coordiImgSrc: "/images/coordiImg/userCoordi4.jpegtype=m_webp",
+    coordiImgType: "jpegtype=m_webp"
+  },
+  {
+    id: 5,
+    thumbnailSrc: "/images/coordiImg/user_chaechae1.jpeg",
+    thumbnailType: "jpeg",
+    userId: "_chaechae_1",
+    coordiImgSrc: "/images/coordiImg/userCoordi5.jpegtype=m_webp",
+    coordiImgType: "jpegtype=m_webp"
+  },
+  {
+    id: 6,
+    thumbnailSrc: "/images/coordiImg/user_actjunho.jpeg",
+    thumbnailType: "jpeg",
+    userId: "act_junho",
+    coordiImgSrc: "/images/coordiImg/userCoordi6.jpegtype=m_webp",
+    coordiImgType: "jpegtype=m_webp"
+  },
+  {
+    id: 7,
+    thumbnailSrc: true,
+    thumbnailType: "jpeg",
+    userId: "lilgaeran",
+    coordiImgSrc: "/images/coordiImg/userCoordi7.jpegtype=m_webp",
+    coordiImgType: "jpegtype=m_webp"
+  },
+  {
+    id: 8,
+    thumbnailSrc: "/images/coordiImg/user_everywear.jpeg",
+    thumbnailType: "jpeg",
+    coordiImgSrc: "/images/coordiImg/userCoordi8.jpegtype=m_webp",
+    coordiImgType: "jpegtype=m_webp"
+  },
+  {
+    id: 9,
+    thumbnailSrc: "/images/coordiImg/user_camperjihyun.jpeg",
+    thumbnailType: "jpeg",
+    userId: "camper_jihyun",
+    coordiImgSrc: "/images/coordiImg/userCoordi9.jpegtype=m_webp",
+    coordiImgType: "jpegtype=m_webp"
+  },
+  {
+    id: 10,
+    thumbnailSrc: "/images/coordiImg/user_blingsso.jpeg",
+    thumbnailType: "jpeg",
+    userId: "bling_sso_",
+    coordiImgSrc: "/images/coordiImg/userCoordi10.jpegtype=m_webp",
+    coordiImgType: "jpegtype=m_webp"
+  },
+  {
+    id: 11,
+    thumbnailSrc: "/images/coordiImg/user_seabrush.jpeg",
+    thumbnailType: "jpeg",
+    userId: "seabrush",
+    coordiImgSrc: "/images/coordiImg/userCoordi11.jpegtype=m_webp",
+    coordiImgType: "jpegtype=m_webp"
+  }
+]
+
+export const newInData= [
+  {
+    id: 0,
+    brand: "ETCE",
+    brandOfficial: true,
+    name: "ETCE Destroyed Padded Bomber Black",
+    price: "352,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/newIn/img0.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 1,
+    brand: "Icyto",
+    brandOfficial: true,
+    name: "Icyto Shield Jeans Black",
+    price: "298,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/newIn/img1.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 2,
+    brand: "STU",
+    brandOfficial: true,
+    name: "STU Women Corduroy Flared Pants Brown",
+    price: "169,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/newIn/img2.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 3,
+    brand: "STU",
+    brandOfficial: true,
+    name: "STU Women Lace Blouse White",
+    price: "192,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/newIn/img3.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 4,
+    brand: "Palace",
+    name: "Palace Basically A Shell 6-Panel Black - 24SS",
+    price: "125,000",
+    saleCount: 1,
+    saleType: false,
+    src: "/images/newIn/img4.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 5,
+    brand: "COTCHS",
+    brandOfficial: true,
+    name: "Cotchs Devon Raglan Tee",
+    price: "65,700",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/newIn/img5.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 6,
+    brand: "OMO",
+    brandOfficial: true,
+    name: "OMO Layered Flared Jean Mud Blue",
+    price: "219,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/newIn/img6.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 7,
+    brand: "COTCHS",
+    brandOfficial: true,
+    name: "Cotchs Brain Hooded Zip Up",
+    price: "121,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/newIn/img7.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 8,
+    brand: "Thevinylhouse",
+    brandOfficial: true,
+    name: "[KREAM Exclusive] Thevinylhouse Bully Keyring Black",
+    price: "39,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/newIn/img8.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 9,
+    brand: "Surgery",
+    brandOfficial: true,
+    name: "Surgery Process 024",
+    price: "1,128,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/newIn/img9.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 10,
+    brand: "PRTPRTSHOP",
+    brandOfficial: true,
+    name: "Sunflower Cayo Shirts White",
+    price: "221,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/newIn/img10.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 11,
+    brand: "OY",
+    brandOfficial: true,
+    name: "OY Layered Hoodie Jacket Dark Grey",
+    price: "215,100",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/newIn/img11.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 12,
+    brand: "ETCE",
+    brandOfficial: true,
+    name: "ETCE Corduroy Raglan Jacket Brown",
+    price: "325,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/newIn/img12.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 13,
+    brand: "Kiko Kostadinov",
+    name: "Kiko Kostadinov Sorelle Jumper Blue",
+    price: "269,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/newIn/img13.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 14,
+    brand: "47 Brand",
+    name: "47 Brand New York Yankees BW Clean Up Black",
+    price: "27,000",
+    saleCount: 61,
+    saleType: false,
+    src: "/images/newIn/img14.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+]
+
+export const weeklyPickData= [
+  {
+    id: 0,
+    brand: "Ottolinger",
+    brandOfficial: true,
+    name: "Ottolinger Women Woven Denim Drape Pants Blue",
+    price: "382,500",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/weeklyPick/img0.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 1,
+    brand: "Woolrich",
+    brandOfficial: true,
+    name: "Woolrich Gtx 2L Fishing Gear Jacket Beige",
+    price: "393,500",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/weeklyPick/img1.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 2,
+    brand: "Ottolinger",
+    brandOfficial: true,
+    name: "Ottolinger Women Signature Split Blazer Black Pinstripe",
+    price: "1,300,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/weeklyPick/img2.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 3,
+    brand: "The Garment",
+    brandOfficial: true,
+    name: "The Garment Women Valetta Shirt Blouse Vanilla",
+    price: "460,700",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/weeklyPick/img3.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 4,
+    brand: "Woolrich",
+    brandOfficial: true,
+    name: "Woolrich Gtx 3L Fishing Gear Jacket Olive",
+    price: "458,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/weeklyPick/img4.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 5,
+    brand: "Werkstatt:Munchen",
+    brandOfficial: true,
+    name: "Werkstatt:Munchen Necklace Two Chains Sterling Silver",
+    price: "545,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/weeklyPick/img5.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 6,
+    brand: "Paula Canovas Del Vas",
+    brandOfficial: true,
+    name: "Paula Canovas Del Vas Women Diablo Heel Sandals Black",
+    price: "565,200",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/weeklyPick/img6.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 7,
+    brand: "Perverze",
+    brandOfficial: true,
+    name: "Perverze Wool String Top and Scarf Red",
+    price: "204,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/weeklyPick/img7.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 8,
+    brand: "Ottolinger",
+    brandOfficial: true,
+    name: "Ottolinger Women Woven Double Waistband Suit Trousers Blue",
+    price: "550,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/weeklyPick/img8.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 9,
+    brand: "AMOMENTO",
+    brandOfficial: true,
+    name: "Amomento Mens Boucle Raglan Coat Brown",
+    price: "538,300",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/weeklyPick/img9.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 10,
+    brand: "Hype Clothing",
+    brandOfficial: true,
+    name: "Nike x Supreme Air Zoom Courtposite SP White University Red",
+    price: "328,500",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/weeklyPick/img10.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 11,
+    brand: "Werkstatt:Munchen",
+    brandOfficial: true,
+    name: "Werkstatt:Munchen 4 Ring Combination Cross Sterling Silver",
+    price: "695,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/weeklyPick/img11.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 12,
+    brand: "Perverze",
+    brandOfficial: true,
+    name: "Perverze Narrow Tie T-Shiet White",
+    price: "150,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/weeklyPick/img12.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 13,
+    brand: "Yunse",
+    brandOfficial: true,
+    name: "Yunse Women Unbalance Pleats Skirt Black",
+    price: "242,200",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/weeklyPick/img13.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 14,
+    brand: "Hoka",
+    brandOfficial: true,
+    name: "Hoka Kaha 2 Low GTX Oxford Tan Dune",
+    price: "279,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/weeklyPick/img14.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+]
+
+export const honeyScoopHouseData = [
+  {
+    id: 0,
+    brand: "Honey Scoop House",
+    brandOfficial: true,
+    name: "Honey Scoop House OE Logo Check LS Shirt Green",
+    price: "79,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/honeyScoopHouse/img0.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 1,
+    brand: "Honey Scoop House",
+    brandOfficial: true,
+    name: "Honey Scoop House OE Logo Check LS Shirt Beige",
+    price: "79,500",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/honeyScoopHouse/img1.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 2,
+    brand: "Honey Scoop House",
+    brandOfficial: true,
+    name: "Honey Scoop House Honey Bear Work Jacket Green",
+    price: "129,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/honeyScoopHouse/img2.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 3,
+    brand: "Honey Scoop House",
+    brandOfficial: true,
+    name: "Honey Scoop House Honey Bear Work Jacket Navy",
+    price: "129,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/honeyScoopHouse/img3.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 4,
+    brand: "Honey Scoop House",
+    brandOfficial: true,
+    name: "Honey Scoop House HNY Logo Trucker Cap Black",
+    price: "55,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/honeyScoopHouse/img4.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 5,
+    brand: "Honey Scoop House",
+    brandOfficial: true,
+    name: "Honey Scoop House BR Logo Trucker Cap Orange",
+    price: "55,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/honeyScoopHouse/img5.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 6,
+    brand: "Honey Scoop House",
+    brandOfficial: true,
+    name: "Honey Scoop House Lips Trucker Cap Royal",
+    price: "55,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/honeyScoopHouse/img6.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 7,
+    brand: "Honey Scoop House",
+    brandOfficial: true,
+    name: "Honey Scoop House Honey Bear Trucker Cap Black",
+    price: "55,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/honeyScoopHouse/img7.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 8,
+    brand: "Honey Scoop House",
+    brandOfficial: true,
+    name: "Honey Scoop House HSH Logo Rug Multi",
+    price: "55,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/honeyScoopHouse/img8.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 9,
+    brand: "Honey Scoop House",
+    brandOfficial: true,
+    name: "Honey Scoop House Clover Logo LSDeep Green",
+    price: "55,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/honeyScoopHouse/img9.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 10,
+    brand: "Honey Scoop House",
+    brandOfficial: true,
+    name: "Honey Scoop House BHF Logo Hooded Sweatshirt Blue",
+    price: "95,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/honeyScoopHouse/img10.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 11,
+    brand: "Honey Scoop House",
+    brandOfficial: true,
+    name: "Honey Scoop House HSH Logo 5-Panel Cap Black",
+    price: "52,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/honeyScoopHouse/img11.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 12,
+    brand: "Honey Scoop House",
+    brandOfficial: true,
+    name: "Honey Scoop House Clover Logo LSBurgundy",
+    price: "55,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/honeyScoopHouse/img12.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 13,
+    brand: "Honey Scoop House",
+    brandOfficial: true,
+    name: "Honey Scoop House BHF Logo Hooded Sweatshirt Black",
+    price: "95,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/honeyScoopHouse/img13.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 14,
+    brand: "Honey Scoop House",
+    brandOfficial: true,
+    name: "Honey Scoop House HSH Logo Beanie Black",
+    price: "32,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/honeyScoopHouse/img14.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+]
+
+export const northFaceData = [
+  {
+    id: 0,
+    brand: "The North Face",
+    brandOfficial: false,
+    name: "The North Face White Label Kulane Down Bomber Black",
+    price: "329,000",
+    saleCount: 11,
+    saleType: true,
+    src: "/images/northFace/img0.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 1,
+    brand: "The North Face",
+    brandOfficial: false,
+    name: "(W) The North Face Greenland Windstopper Jacket Burgandy",
+    price: "240,000",
+    saleCount: 0,
+    saleType: true,
+    src: "/images/northFace/img1.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 2,
+    brand: "The North Face",
+    brandOfficial: false,
+    name: "The North Face Play Green Rvs Fleece Anorak Ivory",
+    price: "90,000",
+    saleCount: 1,
+    saleType: true,
+    src: "/images/northFace/img2.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 3,
+    brand: "The North Face",
+    brandOfficial: false,
+    name: "(W) The North Face Furry Fleece Jacket Black",
+    price: "115,000",
+    saleCount: 0,
+    saleType: true,
+    src: "/images/northFace/img3.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 4,
+    brand: "The North Face",
+    brandOfficial: false,
+    name: "The North Face Denali Wind Jacket Khaki",
+    price: "175,000",
+    saleCount: 1,
+    saleType: true,
+    src: "/images/northFace/img4.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 5,
+    brand: "The North Face",
+    brandOfficial: false,
+    name: "The North Face All-Round Bomber White Sand",
+    price: "155,000",
+    saleCount: 0,
+    saleType: true,
+    src: "/images/northFace/img5.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 6,
+    brand: "The North Face",
+    brandOfficial: false,
+    name: "The North Face Flyhigh Jacket Ink",
+    price: "90,000",
+    saleCount: 0,
+    saleType: true,
+    src: "/images/northFace/img6.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 7,
+    brand: "The North Face",
+    brandOfficial: false,
+    name: "The North Face White Label Neilton Jacket White Sand",
+    price: "169,000",
+    saleCount: 89,
+    saleType: true,
+    src: "/images/northFace/img7.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 8,
+    brand: "The North Face",
+    brandOfficial: false,
+    name: "The North Face Ice Run Jacket Black",
+    price: "168,000",
+    saleCount: 196,
+    saleType: true,
+    src: "/images/northFace/img8.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 9,
+    brand: "The North Face",
+    brandOfficial: false,
+    name: "(W) The North Face White Label Neilton Jacket Brown",
+    price: "144,000",
+    saleCount: 35,
+    saleType: true,
+    src: "/images/northFace/img9.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 10,
+    brand: "The North Face",
+    brandOfficial: false,
+    name: "The North Face Wax Heat Coach Jacket Dark Olive",
+    price: "219,000",
+    saleCount: 3,
+    saleType: true,
+    src: "/images/northFace/img10.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 11,
+    brand: "The North Face",
+    brandOfficial: false,
+    name: "The North Face Novelty GTX Mountain Jacket S Real Black",
+    price: "206,000",
+    saleCount: 3,
+    saleType: true,
+    src: "/images/northFace/img11.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 12,
+    brand: "The North Face",
+    brandOfficial: false,
+    name: "The North Face White Label Comfy Alpha Fleece Zip-Up Light Gray",
+    price: "119,000",
+    saleCount: 85,
+    saleType: true,
+    src: "/images/northFace/img12.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 13,
+    brand: "The North Face",
+    brandOfficial: false,
+    name: "The North Face White Label Rimo Reversible Fleece Hoodie Beige",
+    price: "300,000",
+    saleCount: 33,
+    saleType: true,
+    src: "/images/northFace/img13.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 14,
+    brand: "The North Face",
+    brandOfficial: false,
+    name: "(W) The North Face White Label Neilton Jacket Black",
+    price: "143,000",
+    saleCount: 13,
+    saleType: true,
+    src: "/images/northFace/img14.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+]
+
+export const nylonBagsData = [
+  {
+    id: 0,
+    brand: "Original Sports",
+    brandOfficial: true,
+    name: "Original Sports Nylon String Bag White",
+    price: "80,100",
+    saleCount: 11,
+    saleType: false,
+    src: "/images/nylonBags/img0.jpeg",
+    imgType: "jpeg",
+    bookmarkState: false
+  },
+  {
+    id: 1,
+    brand: "Balenciaga",
+    brandOfficial: false,
+    name: "Balenciaga Recycled Nylon Army Medium Tote Bag Black",
+    price: "1,130,000",
+    saleCount: 4,
+    saleType: true,
+    src: "/images/nylonBags/img1.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 2,
+    brand: "Supreme",
+    brandOfficial: false,
+    name: "Supreme Shoulder Bag Red - 22FW",
+    price: "186,000",
+    saleCount: 15,
+    saleType: true,
+    src: "/images/nylonBags/img2.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 3,
+    brand: "Icyto",
+    brandOfficial: true,
+    name: "Icyto Logo String Backpack White",
+    price: "99,000",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/nylonBags/img3.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 4,
+    brand: "Prada",
+    brandOfficial: false,
+    name: "Prada Re-Nylon and Saffiano Leather Duffle Bag Black",
+    price: "2,199,000",
+    saleCount: 5,
+    saleType: true,
+    src: "/images/nylonBags/img4.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 5,
+    brand: "Notabag",
+    brandOfficial: false,
+    name: "Notabag Original Navy Blue",
+    price: "25,000",
+    saleCount: 0,
+    saleType: true,
+    src: "/images/nylonBags/img5.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 6,
+    brand: "Adidas",
+    brandOfficial: false,
+    name: "Adidas Satin Mini Bucket Backpack Black",
+    price: "99,000",
+    saleCount: 77,
+    saleType: true,
+    src: "/images/nylonBags/img6.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 7,
+    brand: "Bottega Veneta",
+    brandOfficial: false,
+    name: "Bottega Veneta Padded Tech Cassette Frost",
+    price: "2,100,000",
+    saleCount: 15,
+    saleType: true,
+    src: "/images/nylonBags/img7.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 8,
+    brand: "COS",
+    brandOfficial: false,
+    name: "COS Mini Crossbody Pouch Nylon Black",
+    price: "51,000",
+    saleCount: 250,
+    saleType: true,
+    src: "/images/nylonBags/img8.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 9,
+    brand: "Vivienne Westwood",
+    brandOfficial: false,
+    name: "Vivienne Westwood Re-Nylon Drawstring Pouch Set Shoulder Bag Black",
+    price: "364,000",
+    saleCount: 27,
+    saleType: true,
+    src: "/images/nylonBags/img9.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 10,
+    brand: "Adidas",
+    brandOfficial: false,
+    name: "Adidas Adicolor Classic Festival Bag Blue Bird",
+    price: "25,000",
+    saleCount: 0,
+    saleType: true,
+    src: "/images/nylonBags/img10.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 11,
+    brand: "Stone Island",
+    brandOfficial: false,
+    name: "Stone Island 90870 Mussola Gommata Canvas Small Pouch Bag Lead - 23FW",
+    price: "298,000",
+    saleCount: 3,
+    saleType: true,
+    src: "/images/nylonBags/img11.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 12,
+    brand: "Bally",
+    brandOfficial: false,
+    name: "Bally Ferey Nylon Backpack Black",
+    price: "589,000",
+    saleCount: 2,
+    saleType: true,
+    src: "/images/nylonBags/img12.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 13,
+    brand: "Gucci",
+    brandOfficial: false,
+    name: "Gucci Micro GG Ssima Canvas Messenger Bag Black",
+    price: "709,000",
+    saleCount: 6,
+    saleType: true,
+    src: "/images/nylonBags/img13.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 14,
+    brand: "Porter",
+    brandOfficial: false,
+    name: "Porter x Comme des Garcons Backpack L Black",
+    price: "429,000",
+    saleCount: 68,
+    saleType: true,
+    src: "/images/nylonBags/img14.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+]
+
+export const oyData= [
+  {
+    id: 0,
+    brand: "OY",
+    brandOfficial: true,
+    name: "OY S.R Biker Jacket Black",
+    price: "179,100",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/oy/img0.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 1,
+    brand: "OY",
+    brandOfficial: true,
+    name: "OY W.S.L Track Jacket Black",
+    price: "152,100",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/oy/img1.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 2,
+    brand: "OY",
+    brandOfficial: true,
+    name: "OY Nylon Wind Breaker Jacket Black",
+    price: "134,100",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/oy/img2.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 3,
+    brand: "OY",
+    brandOfficial: true,
+    name: "OY Vintage Vegan Leather Hoodie Trucker Jacket Black",
+    price: "197,100",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/oy/img3.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 4,
+    brand: "OY",
+    brandOfficial: true,
+    name: "OY Sporty Opyd Hoodie L.S T Black",
+    price: "107,100",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/oy/img4.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 5,
+    brand: "OY",
+    brandOfficial: true,
+    name: "OY Layered Hoodie Jacket Black",
+    price: "215,100",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/oy/img5.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 6,
+    brand: "OY",
+    brandOfficial: true,
+    name: "OY Heart Stitch Destroyed Knit Black",
+    price: "107,100",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/oy/img6.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 7,
+    brand: "OY",
+    brandOfficial: true,
+    name: "OY Layered Check Shirts Hoodie Black",
+    price: "134,100",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/oy/img7.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 8,
+    brand: "OY",
+    brandOfficial: true,
+    name: "OY Fourteen Cargo Pockets Washed Jeans Black",
+    price: "178,200",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/oy/img8.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 9,
+    brand: "OY",
+    brandOfficial: true,
+    name: "OY W.S.L Tape Track Pants Black",
+    price: "134,100",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/oy/img9.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 10,
+    brand: "OY",
+    brandOfficial: true,
+    name: "OY C.P Odolly Cap Black",
+    price: "44,100",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/oy/img10.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 11,
+    brand: "OY",
+    brandOfficial: true,
+    name: "OY Metal Logo Beanie Black",
+    price: "32,400",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/oy/img11.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 12,
+    brand: "OY",
+    brandOfficial: true,
+    name: "OY S.L Beanie Paul Blue",
+    price: "26,100",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/oy/img12.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 13,
+    brand: "OY",
+    brandOfficial: true,
+    name: "OY Biker Cube Keychain",
+    price: "53,100",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/oy/img13.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 14,
+    brand: "OY",
+    brandOfficial: true,
+    name: "OY Logo Keyring",
+    price: "26,100",
+    saleCount: 0,
+    saleType: false,
+    src: "/images/oy/img14.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+]
+
+export const hotTrandsData = [
+  {
+    id: 0,
+    thumbnailSrc: "/images/hotTrands/user_hiddenbrand.jpeg",
+    thumbnailType: "jpeg",
+    userId: "hiddenbrand",
+    coordiImgSrc: "/images/hotTrands/userTrand0.jpegtype=m_webp",
+    coordiImgType: "jpegtype=m_webp"
+  },
+  {
+    id: 1,
+    thumbnailSrc: "/images/hotTrands/user_sonminsu.png",
+    thumbnailType: "png",
+    userId: "sonminsu",
+    coordiImgSrc: "/images/hotTrands/userTrand1.jpgtype=s2",
+    coordiImgType: "jpgtype=s2"
+  },
+  {
+    id: 2,
+    thumbnailSrc: "/images/hotTrands/user_coordifolder.png",
+    thumbnailType: "png",
+    userId: "CoordiFolder",
+    coordiImgSrc: "/images/hotTrands/userTrand2.jpegtype=m_webp",
+    coordiImgType: "jpegtype=m_webp"
+  },
+  {
+    id: 3,
+    thumbnailSrc: "/images/hotTrands/user_fashionbook.png",
+    thumbnailType: "png",
+    userId: "fashionbook",
+    coordiImgSrc: "/images/hotTrands/userTrand3.jpegtype=m_webp",
+    coordiImgType: "jpegtype=m_webp"
+  },
+  {
+    id: 4,
+    thumbnailSrc: "/images/hotTrands/user_coordifolder.png",
+    thumbnailType: "png",
+    userId: "CoordiFolder",
+    coordiImgSrc: "/images/hotTrands/userTrand4.jpegtype=m_webp",
+    coordiImgType: "jpegtype=m_webp"
+  },
+  {
+    id: 5,
+    thumbnailSrc: "/images/hotTrands/user_everywear.jpeg",
+    thumbnailType: "jpeg",
+    userId: "everywear",
+    coordiImgSrc: "/images/hotTrands/userTrand5.jpgtype=s2",
+    coordiImgType: "jpgtype=s2"
+  },
+  {
+    id: 6,
+    thumbnailSrc: "/images/hotTrands/user_fashionbook.png",
+    thumbnailType: "png",
+    userId: "fashionbook",
+    coordiImgSrc: "/images/hotTrands/userTrand6.jpegtype=m_webp",
+    coordiImgType: "jpegtype=m_webp"
+  },
+  {
+    id: 7,
+    thumbnailSrc: "/images/hotTrands/user_hiddenbrand.jpeg",
+    thumbnailType: "jpeg",
+    userId: "hiddenbrand",
+    coordiImgSrc: "/images/hotTrands/userTrand7.jpegtype=m_webp",
+    coordiImgType: "jpegtype=m_webp"
+  },
+  {
+    id: 8,
+    thumbnailSrc: "/images/hotTrands/user_coordifolder.png",
+    thumbnailType: "png",
+    userId: "CoordiFolder",
+    coordiImgSrc: "/images/hotTrands/userTrand8.jpegtype=m_webp",
+    coordiImgType: "jpegtype=m_webp"
+  },
+  {
+    id: 9,
+    thumbnailSrc: "/images/hotTrands/user_fashionbook.png",
+    thumbnailType: "png",
+    userId: "fashionbook",
+    coordiImgSrc: "/images/hotTrands/userTrand9.jpegtype=m_webp",
+    coordiImgType: "jpegtype=m_webp"
+  },
+  {
+    id: 10,
+    thumbnailSrc: "/images/hotTrands/user_everywear.jpeg",
+    userId: "everywear",
+    coordiImgSrc: "/images/hotTrands/userTrand10.jpgtype=s2",
+    coordiImgType: "jpgtype=s2"
+  }
+]
+
+export const newLowestAsksData= [
+  {
+    id: 0,
+    brand: "Nike",
+    brandOfficial: false,
+    name: "Nike Pro Closed Patella Knee Sleeve 3.0 Black",
+    price: "30,000",
+    saleCount: 10,
+    saleType: true,
+    src: "/images/newLowestAsks/img0.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 1,
+    brand: "Jordan",
+    brandOfficial: false,
+    name: "Jordan 1 x Travis Scott Retro Low OG SP Sail and Ridgerock",
+    price: "1,327,000",
+    saleCount: 145000,
+    saleType: true,
+    src: "/images/newLowestAsks/img1.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 2,
+    brand: "Nike",
+    brandOfficial: false,
+    name: "Nike NSW Club French Terry Crew Sweatshirt Black - Asia",
+    price: "46,000",
+    saleCount: 966,
+    saleType: true,
+    src: "/images/newLowestAsks/img2.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 3,
+    brand: "Nike",
+    brandOfficial: false,
+    name: "Nike ACG Polartec Wolf Tree Pullover Hoodie Anthracite - Asia",
+    price: "150,000",
+    saleCount: 250,
+    saleType: true,
+    src: "/images/newLowestAsks/img3.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 4,
+    brand: "Nike",
+    brandOfficial: false,
+    name: "Nike Air Kukini Worn Blue Obsidian",
+    price: "90,000",
+    saleCount: 422,
+    saleType: true,
+    src: "/images/newLowestAsks/img4.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 5,
+    brand: "Nike",
+    brandOfficial: false,
+    name: "Nike Kobe 4 Protro Black University Gold",
+    price: "249,000",
+    saleCount: 426,
+    saleType: true,
+    src: "/images/newLowestAsks/img5.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 6,
+    brand: "Nike",
+    brandOfficial: false,
+    name: "(W) Nike Daybreak Summit White Black",
+    price: "44,000",
+    saleCount: 6697,
+    saleType: true,
+    src: "/images/newLowestAsks/img6.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 7,
+    brand: "Supreme",
+    brandOfficial: false,
+    name: "Supreme Sleeve Arc Hooded Sweatshirt Heather Grey - 23FW",
+    price: "339,000",
+    saleCount: 79,
+    saleType: true,
+    src: "/images/newLowestAsks/img7.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 8,
+    brand: "Arc'teryx",
+    brandOfficial: false,
+    name: "Arc'teryx Heliad 10 Backpack Forage",
+    price: "300,000",
+    saleCount: 516,
+    saleType: true,
+    src: "/images/newLowestAsks/img8.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 9,
+    brand: "Asics",
+    brandOfficial: false,
+    name: "Asics x Kith Gel-Kayano 14 Cream Scarab",
+    price: "379,000",
+    saleCount: 168,
+    saleType: true,
+    src: "/images/newLowestAsks/img9.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+]
+
+export const newHighestBidsData= [
+  {
+    id: 0,
+    brand: "Nike",
+    brandOfficial: false,
+    name: "Nike Air More Uptempo 96 Copy and Paste",
+    price: "190,000",
+    saleCount: 643,
+    saleType: true,
+    src: "/images/newHighestBids/img0.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 1,
+    brand: "New Balance",
+    brandOfficial: false,
+    name: "New Balance x Aime Leon Dore 550 Silver Mix",
+    price: "200,000",
+    saleCount: 109,
+    saleType: true,
+    src: "/images/newHighestBids/img1.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 2,
+    brand: "Onitsuka Tiger",
+    brandOfficial: false,
+    name: "Onitsuka Tiger Mexico 66 SD Cream Birch",
+    price: "200,000",
+    saleCount: 454,
+    saleType: true,
+    src: "/images/newHighestBids/img2.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 3,
+    brand: "Nike",
+    brandOfficial: false,
+    name: "(W) Nike Court Vision Alta White",
+    price: "52,000",
+    saleCount: 898,
+    saleType: true,
+    src: "/images/newHighestBids/img3.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 4,
+    brand: "New Balance",
+    brandOfficial: false,
+    name: "New Balance 2002R Premium Vintage Navy Grey",
+    price: "170,000",
+    saleCount: 366,
+    saleType: true,
+    src: "/images/newHighestBids/img4.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 5,
+    brand: "Supreme",
+    brandOfficial: false,
+    name: "Supreme Pilled Sweater Black - 23FW",
+    price: "160,000",
+    saleCount: 5,
+    saleType: true,
+    src: "/images/newHighestBids/img5.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 6,
+    brand: "New Balance",
+    brandOfficial: false,
+    name: "New Balance 990v6 Made in USA Sulphur Green",
+    price: "210,000",
+    saleCount: 42,
+    saleType: true,
+    src: "/images/newHighestBids/img6.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 7,
+    brand: "Yeezy",
+    brandOfficial: false,
+    name: "Yeezy Gap Engineered By Balenciaga Dove Hoodie Beige",
+    price: "360,000",
+    saleCount: 288,
+    saleType: true,
+    src: "/images/newHighestBids/img7.jpg",
+    imgType: "jpg",
+    bookmarkState: false
+  },
+  {
+    id: 8,
+    brand: "Adidas",
+    brandOfficial: false,
+    name: "(W) Adidas Spezial Handball Core Black Clear Pink",
+    price: "170,000",
+    saleCount: 2211,
+    saleType: true,
+    src: "/images/newHighestBids/img8.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+  {
+    id: 9,
+    brand: "New Balance",
+    brandOfficial: false,
+    name: "New Balance x J.Crew 2002R Black Powder Green",
+    price: "90,000",
+    saleCount: 223,
+    saleType: true,
+    src: "/images/newHighestBids/img9.png",
+    imgType: "png",
+    bookmarkState: false
+  },
+]
+
+export const upcomingReleaseData= [
+  {
+    id: 0,
+    brand: "Adidas",
+    name: "Adidas Gazelle Indoor Cloud White Collegiate Green",
+    src: "/images/upcomingRelease/img0.png",
+    imgType: "png",
+    upcomingDate: 1
+  },
+  {
+    id: 1,
+    brand: "New Balance",
+    name: "New Balance x JJJJound 2002R Gore-Tex Pine",
+    src: "/images/upcomingRelease/img1.png",
+    imgType: "png",
+    upcomingDate: 1
+  },
+  {
+    id: 2,
+    brand: "New Balance",
+    name: "New Balance x JJJJound 2002R Gore-Tex Charcoal",
+    src: "/images/upcomingRelease/img2.png",
+    imgType: "png",
+    upcomingDate: 1
+  },
+  {
+    id: 3,
+    brand: "Nike",
+    name: "Nike Air Zoom Alphafly Next% 3 Volt",
+    src: "/images/upcomingRelease/img3.png",
+    imgType: "png",
+    upcomingDate: 1
+  },
+  {
+    id: 4,
+    brand: "New Balance",
+    name: "New Balance 998 Made in USA Black Sliver",
+    src: "/images/upcomingRelease/img4.png",
+    imgType: "png",
+    upcomingDate: 1
+  },
+  {
+    id: 5,
+    brand: "Asics",
+    name: "Asics Gel-Kayano 14 Black Pure Silver",
+    src: "/images/upcomingRelease/img5.png",
+    imgType: "png",
+    upcomingDate: 1
+  },
+  {
+    id: 6,
+    brand: "Adidas",
+    name: "Adidas Adistar Cushion Cloud White Silver Metallic",
+    src: "/images/upcomingRelease/img6.png",
+    imgType: "png",
+    upcomingDate: 1
+  },
+  {
+    id: 7,
+    brand: "Clarks",
+    name: "Clarks x Martine Rose Coming Up Roses Torhill HI Green Snake",
+    src: "/images/upcomingRelease/img7.png",
+    imgType: "png",
+    upcomingDate: 2
+  },
+  {
+    id: 8,
+    brand: "Clarks",
+    name: "Clarks x Martine Rose Coming Up Roses Torhill HI Brown Snake",
+    src: "/images/upcomingRelease/img8.png",
+    imgType: "png",
+    upcomingDate: 2
+  },
+  {
+    id: 9,
+    brand: "New Balance",
+    name: "New Balance WRPD Runner Grey",
+    src: "/images/upcomingRelease/img9.png",
+    imgType: "png",
+    upcomingDate: 2
+  },
+  {
+    id: 10,
+    brand: "Nike",
+    name: "(W) Nike Air Footscape Woven Earth Light British Tan",
+    src: "/images/upcomingRelease/img10.png",
+    imgType: "png",
+    upcomingDate: 2
+  },
+  {
+    id: 11,
+    brand: "Jordan",
+    name: "Jordan 1 High 85 Team Red",
+    src: "/images/upcomingRelease/img11.png",
+    imgType: "png",
+    upcomingDate: 2
+  },
+  {
+    id: 12,
+    brand: "Clarks",
+    name: "Clarks x Martine Rose Coming Up Roses Oxford 2 Green Textile",
+    src: "/images/upcomingRelease/img12.png",
+    imgType: "png",
+    upcomingDate: 2
+  },
+  {
+    id: 13,
+    brand: "Clarks",
+    name: "Clarks x Martine Rose Coming Up Roses Oxford 2 Brown Textile",
+    src: "/images/upcomingRelease/img13.png",
+    imgType: "png",
+    upcomingDate: 2
+  },
+  {
+    id: 14,
+    brand: "New Balance",
+    name: "New Balance WRPD Runner Magnet",
+    src: "/images/upcomingRelease/img14.png",
+    imgType: "png",
+    upcomingDate: 2
+  },
+  {
+    id: 15,
+    brand: "Clarks",
+    name: "Clarks x Martine Rose Coming Up Roses Oxford 2 Black Textile",
+    src: "/images/upcomingRelease/img15.png",
+    imgType: "png",
+    upcomingDate: 2
+  },
+  {
+    id: 16,
+    brand: "Adidas",
+    name: "(C) Adidas x Bad Bunny Response CL Wonder White Clear Granite",
+    src: "/images/upcomingRelease/img16.png",
+    imgType: "png",
+    upcomingDate: 3
+  },
+  {
+    id: 17,
+    brand: "Palace",
+    name: "Palace Horsepower Trucker Black - 24SS",
+    src: "/images/upcomingRelease/img17.png",
+    imgType: "png",
+    upcomingDate: 3
+  },
+  {
+    id: 18,
+    brand: "Jordan",
+    name: "(PS) Jordan 4 Retro Bred Reimagined",
+    src: "/images/upcomingRelease/img18.png",
+    imgType: "png",
+    upcomingDate: 3
+  },
+  {
+    id: 19,
+    brand: "Palace",
+    name: "Palace Teddy Ears Earflapper Black - 24SS",
+    src: "/images/upcomingRelease/img19.png",
+    imgType: "png",
+    upcomingDate: 3
+  },
+]
