--- src/coordinationData.js
@@ -1,99 +1,0 @@
-const coordinationData = [
-  {
-    id: 0,
-    thumbnailSrc: "/images/coordiImg/user_yeon2.jpeg",
-    thumbnailType: "jpeg",
-    userId: "y.e.o.n2",
-    coordiImgSrc: "/images/coordiImg/userCoordi0.jpegtype=m_webp",
-    coordiImgType: "jpegtype=m_webp"
-  },
-  {
-    id: 1,
-    thumbnailSrc: "/images/coordiImg/user_sonyunko.jpeg",
-    thumbnailType: "jpeg",
-    userId: "sonyunko",
-    coordiImgSrc: "/images/coordiImg/userCoordi1.jpegtype=m_webp",
-    coordiImgType: "jpegtype=m_webp"
-  },
-  {
-    id: 2,
-    thumbnailSrc: "/images/coordiImg/user_gogoomee.jpeg",
-    thumbnailType: "jpeg",
-    userId: "gogoomee",
-    coordiImgSrc: "/images/coordiImg/userCoordi2.jpegtype=m_webp",
-    coordiImgType: "jpegtype=m_webp"
-  },
-  {
-    id: 3,
-    thumbnailSrc: "/images/coordiImg/user_everywear.jpeg",
-    thumbnailType: "jpeg",
-    userId: "everywear",
-    coordiImgSrc: "/images/coordiImg/userCoordi3.jpegtype=m_webp",
-    coordiImgType: "jpegtype=m_webp"
-  },
-  {
-    id: 4,
-    thumbnailSrc: "/images/coordiImg/user_heeeseok.jpeg",
-    thumbnailType: "jpeg",
-    userId: "heee.seok",
-    coordiImgSrc: "/images/coordiImg/userCoordi4.jpegtype=m_webp",
-    coordiImgType: "jpegtype=m_webp"
-  },
-  {
-    id: 5,
-    thumbnailSrc: "/images/coordiImg/user_chaechae1.jpeg",
-    thumbnailType: "jpeg",
-    userId: "_chaechae_1",
-    coordiImgSrc: "/images/coordiImg/userCoordi5.jpegtype=m_webp",
-    coordiImgType: "jpegtype=m_webp"
-  },
-  {
-    id: 6,
-    thumbnailSrc: "/images/coordiImg/user_actjunho.jpeg",
-    thumbnailType: "jpeg",
-    userId: "act_junho",
-    coordiImgSrc: "/images/coordiImg/userCoordi6.jpegtype=m_webp",
-    coordiImgType: "jpegtype=m_webp"
-  },
-  {
-    id: 7,
-    thumbnailSrc: true,
-    thumbnailType: "jpeg",
-    userId: "lilgaeran",
-    coordiImgSrc: "/images/coordiImg/userCoordi7.jpegtype=m_webp",
-    coordiImgType: "jpegtype=m_webp"
-  },
-  {
-    id: 8,
-    thumbnailSrc: "/images/coordiImg/user_everywear.jpeg",
-    thumbnailType: "jpeg",
-    coordiImgSrc: "/images/coordiImg/userCoordi8.jpegtype=m_webp",
-    coordiImgType: "jpegtype=m_webp"
-  },
-  {
-    id: 9,
-    thumbnailSrc: "/images/coordiImg/user_camperjihyun.jpeg",
-    thumbnailType: "jpeg",
-    userId: "camper_jihyun",
-    coordiImgSrc: "/images/coordiImg/userCoordi9.jpegtype=m_webp",
-    coordiImgType: "jpegtype=m_webp"
-  },
-  {
-    id: 10,
-    thumbnailSrc: "/images/coordiImg/user_blingsso.jpeg",
-    thumbnailType: "jpeg",
-    userId: "bling_sso_",
-    coordiImgSrc: "/images/coordiImg/userCoordi10.jpegtype=m_webp",
-    coordiImgType: "jpegtype=m_webp"
-  },
-  {
-    id: 11,
-    thumbnailSrc: "/images/coordiImg/user_seabrush.jpeg",
-    thumbnailType: "jpeg",
-    userId: "seabrush",
-    coordiImgSrc: "/images/coordiImg/userCoordi11.jpegtype=m_webp",
-    coordiImgType: "jpegtype=m_webp"
-  }
-]
-
-export default coordinationData;
