--- src/eventBannerData.js
@@ -1,49 +1,0 @@
-const eventBannerData = [
-  {
-    id: 0,
-    src: "/images/eventBanner/banner_0.webp",
-    bgColor: "#6934FF"
-  },
-  {
-    id: 1,
-    src: "/images/eventBanner/banner_1.webp",
-    bgColor: "#6B87A7"
-  },
-  {
-    id: 2,
-    src: "/images/eventBanner/banner_2.webp",
-    bgColor: "#4567E3"
-  },
-  {
-    id: 3,
-    src: "/images/eventBanner/banner_3.webp",
-    bgColor: "#F9E95F"
-  },
-  {
-    id: 4,
-    src: "/images/eventBanner/banner_4.webp",
-    bgColor: "rgb(158, 163, 149)"
-  },
-  {
-    id: 5,
-    src: "/images/eventBanner/banner_5.webp",
-    bgColor: "rgb(110, 127, 119)"
-  },
-  {
-    id: 6,
-    src: "/images/eventBanner/banner_6.webp",
-    bgColor: "rgb(110, 111, 115)"
-  },
-  {
-    id: 7,
-    src: "/images/eventBanner/banner_7.webp",
-    bgColor: "rgb(93, 137, 147)"
-  },
-  {
-    id: 8,
-    src: "/images/eventBanner/banner_8.webp",
-    bgColor: "rgb(138, 131, 131)"
-  },
-]
-
-export default eventBannerData;
