--- src/eventData.js
@@ -1,256 +1,0 @@
-const eventData = [
-  {
-    id: 0,
-    src: "/images/eventBanner/bg_0.webp",
-    bgColor: "#A9A099",
-    activeColor: true
-  },
-  {
-    id: 1,
-    src: "/images/eventBanner/bg_1.webp",
-    bgColor: "#9A9196",
-    activeColor: true
-  },
-  {
-    id: 2,
-    src: "/images/eventBanner/bg_2.webp",
-    bgColor: "#868686",
-    activeColor: true
-  },
-  {
-    id: 3,
-    src: "/images/eventBanner/bg_3.webp",
-    bgColor: "#9796CE",
-    activeColor: true
-  }
-  ,{
-    id: 4,
-    src: "/images/eventBanner/bg_4.webp",
-    bgColor: "#828177",
-    activeColor: true
-  },
-  {
-    id: 5,
-    src: "/images/eventBanner/bg_5.webp",
-    bgColor: "#BABABA",
-    activeColor: true
-  },
-  {
-    id: 6,
-    src: "/images/eventBanner/bg_6.webp",
-    bgColor: "#E5E5E5",
-    activeColor: false
-  },
-  {
-    id: 7,
-    src: "/images/eventBanner/bg_7.webp",
-    bgColor: "#7B7678",
-    activeColor: true
-  },
-  {
-    id: 8,
-    src: "/images/eventBanner/bg_8.webp",
-    bgColor: "#E0FA9D",
-    activeColor: false
-  }
-  ,{
-    id: 9,
-    src: "/images/eventBanner/bg_9.webp",
-    bgColor: "#e5e5e5",
-    activeColor: false
-  },
-  {
-    id: 10,
-    src: "/images/eventBanner/bg_10.webp",
-    bgColor: "#D5DEED",
-    activeColor: false
-  },
-  {
-    id: 11,
-    src: "/images/eventBanner/bg_11.webp",
-    bgColor: "#86807E",
-    activeColor: true
-  },
-  {
-    id: 12,
-    src: "/images/eventBanner/bg_12.webp",
-    bgColor: "#025396",
-    activeColor: true
-  },
-  {
-    id: 13,
-    src: "/images/eventBanner/bg_13.webp",
-    bgColor: "#EFEFEF",
-    activeColor: false
-  }
-  ,{
-    id: 14,
-    src: "/images/eventBanner/bg_14.webp",
-    bgColor: "#030802",
-    activeColor: true
-  },
-  {
-    id: 15,
-    src: "/images/eventBanner/bg_15.webp",
-    bgColor: "#BFAD6E",
-    activeColor: true
-  },
-  {
-    id: 16,
-    src: "/images/eventBanner/bg_16.webp",
-    bgColor: "#DDDBC8",
-    activeColor: true
-  },
-  {
-    id: 17,
-    src: "/images/eventBanner/bg_17.webp",
-    bgColor: "#251313",
-    activeColor: true
-  },
-  {
-    id: 18,
-    src: "/images/eventBanner/bg_18.webp",
-    bgColor: "#0A0A07",
-    activeColor: true
-  }
-  ,{
-    id: 19,
-    src: "/images/eventBanner/bg_19.webp",
-    bgColor: "#0D0D0E",
-    activeColor: true
-  },
-  {
-    id: 20,
-    src: "/images/eventBanner/bg_20.webp",
-    bgColor: "#ACA79C",
-    activeColor: true
-  },
-  {
-    id: 21,
-    src: "/images/eventBanner/bg_21.webp",
-    bgColor: "#198dfa",
-    activeColor: true
-  },
-  {
-    id: 22,
-    src: "/images/eventBanner/bg_22.webp",
-    bgColor: "#252829",
-    activeColor: true
-  },
-  {
-    id: 23,
-    src: "/images/eventBanner/bg_23.webp",
-    bgColor: "#777E82",
-    activeColor: true
-  }
-  ,{
-    id: 24,
-    src: "/images/eventBanner/bg_24.webp",
-    bgColor: "#F9E5EE",
-    activeColor: false
-  },
-  {
-    id: 25,
-    src: "/images/eventBanner/bg_25.webp",
-    bgColor: "#50575A",
-    activeColor: true
-  },
-  {
-    id: 26,
-    src: "/images/eventBanner/bg_26.webp",
-    bgColor: "#C7AE9B",
-    activeColor: true
-  },
-  {
-    id: 27,
-    src: "/images/eventBanner/bg_27.webp",
-    bgColor: "#f5f5f5",
-    activeColor: false
-  },
-  {
-    id: 28,
-    src: "/images/eventBanner/bg_28.webp",
-    bgColor: "#E0E0E0",
-    activeColor: false
-  }
-  ,{
-    id: 29,
-    src: "/images/eventBanner/bg_29.webp",
-    bgColor: "#C8CDC8",
-    activeColor: true
-  },
-  {
-    id: 30,
-    src: "/images/eventBanner/bg_30.webp",
-    bgColor: "#D4E9FA",
-    activeColor: false
-  },
-  {
-    id: 31,
-    src: "/images/eventBanner/bg_31.webp",
-    bgColor: "#57251A",
-    activeColor: true
-  },
-  {
-    id: 32,
-    src: "/images/eventBanner/bg_32.webp",
-    bgColor: "#2BB165",
-    activeColor: true
-  },
-  {
-    id: 33,
-    src: "/images/eventBanner/bg_33.webp",
-    bgColor: "#000000",
-    activeColor: true
-  }
-  ,{
-    id: 34,
-    src: "/images/eventBanner/bg_34.webp",
-    bgColor: "#AA9A7C",
-    activeColor: true
-  },
-  {
-    id: 35,
-    src: "/images/eventBanner/bg_35.webp",
-    bgColor: "#F5F6F3",
-    activeColor: false
-  },
-  {
-    id: 36,
-    src: "/images/eventBanner/bg_36.webp",
-    bgColor: "#C387AF",
-    activeColor: true
-  },
-  {
-    id: 37,
-    src: "/images/eventBanner/bg_37.webp",
-    bgColor: "#020202",
-    activeColor: true
-  },
-  {
-    id: 38,
-    src: "/images/eventBanner/bg_38.webp",
-    bgColor: "#A9A69D",
-    activeColor: true
-  },
-  {
-    id: 39,
-    src: "/images/eventBanner/bg_39.webp",
-    bgColor: "#d8dee2",
-    activeColor: false
-  },
-  {
-    id: 40,
-    src: "/images/eventBanner/bg_40.webp",
-    bgColor: "#222222",
-    activeColor: true
-  },
-  {
-    id: 41,
-    src: "/images/eventBanner/bg_41.webp",
-    bgColor: "#d8dee2",
-    activeColor: false
-  },
-]
-
-export default eventData;
