--- src/eventDetailData.js
@@ -1,63 +1,0 @@
-const eventDetailData = [
-  {
-    id: 0,
-    type: "webp",
-    src: "/images/eventDetailBtn/bg_0.gif",
-    content: "내 폰 시세"
-  },
-  {
-    id: 1,
-    type: "jpeg",
-    src: "/images/eventDetailBtn/bg_1.jpeg",
-    content: "크림 드로우"
-  },
-  {
-    id: 2,
-    type: "jpg",
-    src: "/images/eventDetailBtn/bg_2.jpg",
-    content: "남성 추천"
-  },
-  {
-    id: 3,
-    type: "jpg",
-    src: "/images/eventDetailBtn/bg_3.jpg",
-    content: "여성 추천"
-  },
-  {
-    id: 4,
-    type: "jpg",
-    src: "/images/eventDetailBtn/bg_4.jpg",
-    content: "색다른 추천"
-  },
-  {
-    id: 5,
-    type: "jpg",
-    src: "/images/eventDetailBtn/bg_5.jpg",
-    content: "정가 아래"
-  },
-  {
-    id: 6,
-    type: "jpg",
-    src: "/images/eventDetailBtn/bg_6.jpg",
-    content: "시즌 오프 ~70%"
-  },
-  {
-    id: 7,
-    type: "jpg",
-    src: "/images/eventDetailBtn/bg_7.jpg",
-    content: "역시즌 최대 82%"
-  },
-  {
-    id: 8,
-    type: "jpg",
-    src: "/images/eventDetailBtn/bg_8.jpg",
-    content: "2월 혜택"
-  },
-  {
-    id: 9,
-    type: "jpg",
-    src: "/images/eventDetailBtn/bg_9.jpg",
-    content: "눕시 & 아식스 선물"
-  }
-]
-export default eventDetailData;
