--- src/honeyScoopHouseData.js
@@ -1,184 +1,0 @@
-const honeyScoopHouseData = [
-  {
-    id: 0,
-    brand: "Honey Scoop House",
-    brandOfficial: true,
-    name: "Honey Scoop House OE Logo Check LS Shirt Green",
-    price: "79,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/honeyScoopHouse/img0.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 1,
-    brand: "Honey Scoop House",
-    brandOfficial: true,
-    name: "Honey Scoop House OE Logo Check LS Shirt Beige",
-    price: "79,500",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/honeyScoopHouse/img1.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 2,
-    brand: "Honey Scoop House",
-    brandOfficial: true,
-    name: "Honey Scoop House Honey Bear Work Jacket Green",
-    price: "129,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/honeyScoopHouse/img2.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 3,
-    brand: "Honey Scoop House",
-    brandOfficial: true,
-    name: "Honey Scoop House Honey Bear Work Jacket Navy",
-    price: "129,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/honeyScoopHouse/img3.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 4,
-    brand: "Honey Scoop House",
-    brandOfficial: true,
-    name: "Honey Scoop House HNY Logo Trucker Cap Black",
-    price: "55,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/honeyScoopHouse/img4.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 5,
-    brand: "Honey Scoop House",
-    brandOfficial: true,
-    name: "Honey Scoop House BR Logo Trucker Cap Orange",
-    price: "55,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/honeyScoopHouse/img5.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 6,
-    brand: "Honey Scoop House",
-    brandOfficial: true,
-    name: "Honey Scoop House Lips Trucker Cap Royal",
-    price: "55,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/honeyScoopHouse/img6.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 7,
-    brand: "Honey Scoop House",
-    brandOfficial: true,
-    name: "Honey Scoop House Honey Bear Trucker Cap Black",
-    price: "55,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/honeyScoopHouse/img7.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 8,
-    brand: "Honey Scoop House",
-    brandOfficial: true,
-    name: "Honey Scoop House HSH Logo Rug Multi",
-    price: "55,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/honeyScoopHouse/img8.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 9,
-    brand: "Honey Scoop House",
-    brandOfficial: true,
-    name: "Honey Scoop House Clover Logo LSDeep Green",
-    price: "55,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/honeyScoopHouse/img9.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 10,
-    brand: "Honey Scoop House",
-    brandOfficial: true,
-    name: "Honey Scoop House BHF Logo Hooded Sweatshirt Blue",
-    price: "95,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/honeyScoopHouse/img10.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 11,
-    brand: "Honey Scoop House",
-    brandOfficial: true,
-    name: "Honey Scoop House HSH Logo 5-Panel Cap Black",
-    price: "52,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/honeyScoopHouse/img11.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 12,
-    brand: "Honey Scoop House",
-    brandOfficial: true,
-    name: "Honey Scoop House Clover Logo LSBurgundy",
-    price: "55,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/honeyScoopHouse/img12.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 13,
-    brand: "Honey Scoop House",
-    brandOfficial: true,
-    name: "Honey Scoop House BHF Logo Hooded Sweatshirt Black",
-    price: "95,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/honeyScoopHouse/img13.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 14,
-    brand: "Honey Scoop House",
-    brandOfficial: true,
-    name: "Honey Scoop House HSH Logo Beanie Black",
-    price: "32,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/honeyScoopHouse/img14.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-]
-
-export default honeyScoopHouseData;
