--- src/hotTrandsData.js
@@ -1,91 +1,0 @@
-const hotTrandsData = [
-  {
-    id: 0,
-    thumbnailSrc: "/images/hotTrands/user_hiddenbrand.jpeg",
-    thumbnailType: "jpeg",
-    userId: "hiddenbrand",
-    coordiImgSrc: "/images/hotTrands/userTrand0.jpegtype=m_webp",
-    coordiImgType: "jpegtype=m_webp"
-  },
-  {
-    id: 1,
-    thumbnailSrc: "/images/hotTrands/user_sonminsu.png",
-    thumbnailType: "png",
-    userId: "sonminsu",
-    coordiImgSrc: "/images/hotTrands/userTrand1.jpgtype=s2",
-    coordiImgType: "jpgtype=s2"
-  },
-  {
-    id: 2,
-    thumbnailSrc: "/images/hotTrands/user_coordifolder.png",
-    thumbnailType: "png",
-    userId: "CoordiFolder",
-    coordiImgSrc: "/images/hotTrands/userTrand2.jpegtype=m_webp",
-    coordiImgType: "jpegtype=m_webp"
-  },
-  {
-    id: 3,
-    thumbnailSrc: "/images/hotTrands/user_fashionbook.png",
-    thumbnailType: "png",
-    userId: "fashionbook",
-    coordiImgSrc: "/images/hotTrands/userTrand3.jpegtype=m_webp",
-    coordiImgType: "jpegtype=m_webp"
-  },
-  {
-    id: 4,
-    thumbnailSrc: "/images/hotTrands/user_coordifolder.png",
-    thumbnailType: "png",
-    userId: "CoordiFolder",
-    coordiImgSrc: "/images/hotTrands/userTrand4.jpegtype=m_webp",
-    coordiImgType: "jpegtype=m_webp"
-  },
-  {
-    id: 5,
-    thumbnailSrc: "/images/hotTrands/user_everywear.jpeg",
-    thumbnailType: "jpeg",
-    userId: "everywear",
-    coordiImgSrc: "/images/hotTrands/userTrand5.jpgtype=s2",
-    coordiImgType: "jpgtype=s2"
-  },
-  {
-    id: 6,
-    thumbnailSrc: "/images/hotTrands/user_fashionbook.png",
-    thumbnailType: "png",
-    userId: "fashionbook",
-    coordiImgSrc: "/images/hotTrands/userTrand6.jpegtype=m_webp",
-    coordiImgType: "jpegtype=m_webp"
-  },
-  {
-    id: 7,
-    thumbnailSrc: "/images/hotTrands/user_hiddenbrand.jpeg",
-    thumbnailType: "jpeg",
-    userId: "hiddenbrand",
-    coordiImgSrc: "/images/hotTrands/userTrand7.jpegtype=m_webp",
-    coordiImgType: "jpegtype=m_webp"
-  },
-  {
-    id: 8,
-    thumbnailSrc: "/images/hotTrands/user_coordifolder.png",
-    thumbnailType: "png",
-    userId: "CoordiFolder",
-    coordiImgSrc: "/images/hotTrands/userTrand8.jpegtype=m_webp",
-    coordiImgType: "jpegtype=m_webp"
-  },
-  {
-    id: 9,
-    thumbnailSrc: "/images/hotTrands/user_fashionbook.png",
-    thumbnailType: "png",
-    userId: "fashionbook",
-    coordiImgSrc: "/images/hotTrands/userTrand9.jpegtype=m_webp",
-    coordiImgType: "jpegtype=m_webp"
-  },
-  {
-    id: 10,
-    thumbnailSrc: "/images/hotTrands/user_everywear.jpeg",
-    userId: "everywear",
-    coordiImgSrc: "/images/hotTrands/userTrand10.jpgtype=s2",
-    coordiImgType: "jpgtype=s2"
-  }
-]
-
-export default hotTrandsData;
