--- src/justDroppedData.js
@@ -1,169 +1,0 @@
-const justDroppedData= [
-  {
-    id: 0,
-    brand: "Vandy The Pink",
-    name: "Vandy The Pink x Alpha Industries Bomber Jacket Lead Gray",
-    price: "489,000",
-    saleCount: 34,
-    saleType: true,
-    src: "/images/justDropped/img0.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 1,
-    brand: "New Balance",
-    name: "New Balance 1906R Lunar New Year Pink",
-    price: "489,000",
-    saleCount: 12,
-    saleType: true,
-    src: "/images/justDropped/img1.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 2,
-    brand: "Palace",
-    name: "Palace Korea Varsity Jacket Black - 24SS",
-    price: "1,080,000",
-    saleCount: 19,
-    saleType: true,
-    src: "/images/justDropped/img2.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 3,
-    brand: "Human Made",
-    name: "Human Made Reversible Yokosuka Jacket Black",
-    price: "1,600,000",
-    saleCount: 1,
-    saleType: true,
-    src: "/images/justDropped/img3.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 4,
-    brand: "Blitzway",
-    name: "[예약 배송] Blitzway Disney D100 Titanium Mickey Mouse",
-    price: "2,899,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/justDropped/img4.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 5,
-    brand: "Asics",
-    name: "Asics Gel-Quantum Kinetic Mid Grey Pure Silver",
-    price: "332,000",
-    saleCount: 27,
-    saleType: true,
-    src: "/images/justDropped/img5.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 6,
-    brand: "Converse",
-    name: "Converse x Feng Chen Wang Chuck 70 Ox 2-in-1 Brown Rice",
-    price: "228,000",
-    saleCount: 63,
-    saleType: true,
-    src: "/images/justDropped/img6.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 7,
-    brand: "Nike",
-    name: "Nike Dunk Low SP Brazil 2024",
-    price: "160,000",
-    saleCount: 31,
-    saleType: true,
-    src: "/images/justDropped/img7.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 8,
-    brand: "Hoka",
-    name: "Hoka Bondi 8 Shifting Sand Vanilla",
-    price: "219,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/justDropped/img8.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 9,
-    brand: "Starbucks",
-    name: "Starbucks Be Mine Bear Fabric Keychain",
-    price: "36,000",
-    saleCount: 113,
-    saleType: true,
-    src: "/images/justDropped/img9.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 10,
-    brand: "New Balance",
-    name: "New Balance 860v2 Optic White",
-    price: "162,000",
-    saleCount: 63,
-    saleType: true,
-    src: "/images/justDropped/img10.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 11,
-    brand: "Adidas",
-    name: "(W) Adidas Samba OG Core Black Wonder White",
-    price: "145,000",
-    saleCount: 201,
-    saleType: true,
-    src: "/images/justDropped/img11.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 12,
-    brand: "Jordan",
-    name: "Jordan 1 Retro High OG Black White",
-    price: "199,000",
-    saleCount: 6,
-    saleType: true,
-    src: "/images/justDropped/img12.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 13,
-    brand: "Nike",
-    name: "Nike x Off-White Allover Print Jersey Kelly Green - Asia",
-    price: "200,000",
-    saleCount: 156,
-    saleType: true,
-    src: "/images/justDropped/img13.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 14,
-    brand: "Stussy",
-    name: "Stussy Wave Dye Beach Shell Jacket Black",
-    price: "323,000",
-    saleCount: 158,
-    saleType: true,
-    src: "/images/justDropped/img14.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-]
-
-export default justDroppedData;
