--- src/mostPopularData.js
@@ -1,227 +1,0 @@
-const mostPopularData= [
-  {
-    id: 0,
-    brand: "Nike",
-    name: "Nike Air Force 1 '07 WB Flax",
-    price: "144,000",
-    saleCount: 147000,
-    saleType: true,
-    src: "/images/mostPopular/img0.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 1,
-    brand: "Jordan",
-    name: "Jordan 4 Retro Bred Reimagined",
-    price: "287,000",
-    saleCount: 236,
-    saleType: true,
-    src: "/images/mostPopular/img1.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 2,
-    brand: "Adidas",
-    name: "Adidas Samba OG Cloud White",
-    price: "117,000",
-    saleCount: 58000,
-    saleType: true,
-    src: "/images/mostPopular/img2.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 3,
-    brand: "Nike",
-    name: "Nike Air Force 1 '07 Low White",
-    price: "123,000",
-    saleCount: 438000,
-    saleType: true,
-    src: "/images/mostPopular/img3.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 4,
-    brand: "Apple",
-    name: "Apple Airpods Pro 2nd Gen with MagSafe Charging Case USB C Type (Korean Ver.)",
-    price: "306,000",
-    saleCount: 6655,
-    saleType: true,
-    src: "/images/mostPopular/img4.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-
-  {
-    id: 5,
-    brand: "Adidas",
-    name: "(W) Adidas Spezial Handball Clear Pink",
-    price: "129,000",
-    saleCount: 6777,
-    saleType: true,
-    src: "/images/mostPopular/img5.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 6,
-    brand: "Adidas",
-    name: "Adidas Samba OG Black White Gum",
-    price: "119,000",
-    saleCount: 52000,
-    saleType: true,
-    src: "/images/mostPopular/img6.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 7,
-    brand: "New Balance",
-    name: "New Balance 993 Made in USA Grey - D Standard",
-    price: "310,000",
-    saleCount: 64000,
-    saleType: true,
-    src: "/images/mostPopular/img7.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 8,
-    brand: "Nike",
-    name: "(W) Nike Zoom Vomero 5 Photon Dust and Metallic Silver",
-    price: "172,000",
-    saleCount: 15000,
-    saleType: true,
-    src: "/images/mostPopular/img8.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 9,
-    brand: "Adidas",
-    name: "(W) Adidas Spezial Handball Core Black Clear Pink",
-    price: "147,000",
-    saleCount: 1406,
-    saleType: true,
-    src: "/images/mostPopular/img9.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-
-  {
-    id: 10,
-    brand: "Mihara Yasuhiro",
-    name: "Maison Mihara Yasuhiro Peterson OG Sole Canvas Low Sneakers Black",
-    price: "250,000",
-    saleCount: 24000,
-    saleType: true,
-    src: "/images/mostPopular/img10.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 11,
-    brand: "Asics",
-    name: "Asics Gel-Kayano 14 Cream Black",
-    price: "224,000",
-    saleCount: 201,
-    saleType: true,
-    src: "/images/mostPopular/img11.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 12,
-    brand: "Nike",
-    name: "Nike Air Force 1 '07 Low Triple Black",
-    price: "124,000",
-    saleCount: 71000,
-    saleType: true,
-    src: "/images/mostPopular/img12.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 13,
-    brand: "Stussy",
-    name: "Stussy Basic Stussy Hoodie Black 2023",
-    price: "222,000",
-    saleCount: 9520,
-    saleType: true,
-    src: "/images/mostPopular/img13.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 14,
-    brand: "Nike",
-    name: "Nike x Ambush Air Force 1 Low SP Black",
-    price: "298,000",
-    saleCount: 7331,
-    saleType: true,
-    src: "/images/mostPopular/img14.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-
-  {
-    id: 15,
-    brand: "Nike",
-    name: "(W) Nike Air Force 1 '07 Low White",
-    price: "110,000",
-    saleCount: 151000,
-    saleType: true,
-    src: "/images/mostPopular/img15.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 16,
-    brand: "New Balance",
-    name: "New Balance 2002R Grey",
-    price: "159,000",
-    saleCount: 66000,
-    saleType: true,
-    src: "/images/mostPopular/img16.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 17,
-    brand: "Nike",
-    name: "Nike Zoom Vomero 5 PRM Light Iron Ore and Flat Fewter",
-    price: "150,000",
-    saleCount: 6,
-    saleType: true,
-    src: "/images/mostPopular/img17.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 18,
-    brand: "Nike",
-    name: "(W) Nike Air Force 1 '07 Cacao Wow",
-    price: "234,000",
-    saleCount: 4018,
-    saleType: true,
-    src: "/images/mostPopular/img18.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 19,
-    brand: "Nike",
-    name: "(W) Nike Zoom Vomero 5 Cobblestone and Flat Pewter",
-    price: "159,000",
-    saleCount: 34000,
-    saleType: true,
-    src: "/images/mostPopular/img19.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-]
-
-export default mostPopularData;
