--- src/newHighestBidsData.js
@@ -1,124 +1,0 @@
-const newHighestBidsData= [
-  {
-    id: 0,
-    brand: "Nike",
-    brandOfficial: false,
-    name: "Nike Air More Uptempo 96 Copy and Paste",
-    price: "190,000",
-    saleCount: 643,
-    saleType: true,
-    src: "/images/newHighestBids/img0.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 1,
-    brand: "New Balance",
-    brandOfficial: false,
-    name: "New Balance x Aime Leon Dore 550 Silver Mix",
-    price: "200,000",
-    saleCount: 109,
-    saleType: true,
-    src: "/images/newHighestBids/img1.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 2,
-    brand: "Onitsuka Tiger",
-    brandOfficial: false,
-    name: "Onitsuka Tiger Mexico 66 SD Cream Birch",
-    price: "200,000",
-    saleCount: 454,
-    saleType: true,
-    src: "/images/newHighestBids/img2.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 3,
-    brand: "Nike",
-    brandOfficial: false,
-    name: "(W) Nike Court Vision Alta White",
-    price: "52,000",
-    saleCount: 898,
-    saleType: true,
-    src: "/images/newHighestBids/img3.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 4,
-    brand: "New Balance",
-    brandOfficial: false,
-    name: "New Balance 2002R Premium Vintage Navy Grey",
-    price: "170,000",
-    saleCount: 366,
-    saleType: true,
-    src: "/images/newHighestBids/img4.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 5,
-    brand: "Supreme",
-    brandOfficial: false,
-    name: "Supreme Pilled Sweater Black - 23FW",
-    price: "160,000",
-    saleCount: 5,
-    saleType: true,
-    src: "/images/newHighestBids/img5.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 6,
-    brand: "New Balance",
-    brandOfficial: false,
-    name: "New Balance 990v6 Made in USA Sulphur Green",
-    price: "210,000",
-    saleCount: 42,
-    saleType: true,
-    src: "/images/newHighestBids/img6.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 7,
-    brand: "Yeezy",
-    brandOfficial: false,
-    name: "Yeezy Gap Engineered By Balenciaga Dove Hoodie Beige",
-    price: "360,000",
-    saleCount: 288,
-    saleType: true,
-    src: "/images/newHighestBids/img7.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 8,
-    brand: "Adidas",
-    brandOfficial: false,
-    name: "(W) Adidas Spezial Handball Core Black Clear Pink",
-    price: "170,000",
-    saleCount: 2211,
-    saleType: true,
-    src: "/images/newHighestBids/img8.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 9,
-    brand: "New Balance",
-    brandOfficial: false,
-    name: "New Balance x J.Crew 2002R Black Powder Green",
-    price: "90,000",
-    saleCount: 223,
-    saleType: true,
-    src: "/images/newHighestBids/img9.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-]
-
-export default newHighestBidsData;
