--- src/newInData.js
@@ -1,181 +1,0 @@
-const newInData= [
-  {
-    id: 0,
-    brand: "ETCE",
-    brandOfficial: true,
-    name: "ETCE Destroyed Padded Bomber Black",
-    price: "352,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/newIn/img0.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 1,
-    brand: "Icyto",
-    brandOfficial: true,
-    name: "Icyto Shield Jeans Black",
-    price: "298,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/newIn/img1.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 2,
-    brand: "STU",
-    brandOfficial: true,
-    name: "STU Women Corduroy Flared Pants Brown",
-    price: "169,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/newIn/img2.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 3,
-    brand: "STU",
-    brandOfficial: true,
-    name: "STU Women Lace Blouse White",
-    price: "192,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/newIn/img3.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 4,
-    brand: "Palace",
-    name: "Palace Basically A Shell 6-Panel Black - 24SS",
-    price: "125,000",
-    saleCount: 1,
-    saleType: false,
-    src: "/images/newIn/img4.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 5,
-    brand: "COTCHS",
-    brandOfficial: true,
-    name: "Cotchs Devon Raglan Tee",
-    price: "65,700",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/newIn/img5.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 6,
-    brand: "OMO",
-    brandOfficial: true,
-    name: "OMO Layered Flared Jean Mud Blue",
-    price: "219,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/newIn/img6.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 7,
-    brand: "COTCHS",
-    brandOfficial: true,
-    name: "Cotchs Brain Hooded Zip Up",
-    price: "121,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/newIn/img7.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 8,
-    brand: "Thevinylhouse",
-    brandOfficial: true,
-    name: "[KREAM Exclusive] Thevinylhouse Bully Keyring Black",
-    price: "39,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/newIn/img8.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 9,
-    brand: "Surgery",
-    brandOfficial: true,
-    name: "Surgery Process 024",
-    price: "1,128,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/newIn/img9.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 10,
-    brand: "PRTPRTSHOP",
-    brandOfficial: true,
-    name: "Sunflower Cayo Shirts White",
-    price: "221,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/newIn/img10.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 11,
-    brand: "OY",
-    brandOfficial: true,
-    name: "OY Layered Hoodie Jacket Dark Grey",
-    price: "215,100",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/newIn/img11.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 12,
-    brand: "ETCE",
-    brandOfficial: true,
-    name: "ETCE Corduroy Raglan Jacket Brown",
-    price: "325,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/newIn/img12.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 13,
-    brand: "Kiko Kostadinov",
-    name: "Kiko Kostadinov Sorelle Jumper Blue",
-    price: "269,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/newIn/img13.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 14,
-    brand: "47 Brand",
-    name: "47 Brand New York Yankees BW Clean Up Black",
-    price: "27,000",
-    saleCount: 61,
-    saleType: false,
-    src: "/images/newIn/img14.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-]
-
-export default newInData;
