--- src/newLowestAsksData.js
@@ -1,124 +1,0 @@
-const newLowestAsksData= [
-  {
-    id: 0,
-    brand: "Nike",
-    brandOfficial: false,
-    name: "Nike Pro Closed Patella Knee Sleeve 3.0 Black",
-    price: "30,000",
-    saleCount: 10,
-    saleType: true,
-    src: "/images/newLowestAsks/img0.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 1,
-    brand: "Jordan",
-    brandOfficial: false,
-    name: "Jordan 1 x Travis Scott Retro Low OG SP Sail and Ridgerock",
-    price: "1,327,000",
-    saleCount: 145000,
-    saleType: true,
-    src: "/images/newLowestAsks/img1.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 2,
-    brand: "Nike",
-    brandOfficial: false,
-    name: "Nike NSW Club French Terry Crew Sweatshirt Black - Asia",
-    price: "46,000",
-    saleCount: 966,
-    saleType: true,
-    src: "/images/newLowestAsks/img2.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 3,
-    brand: "Nike",
-    brandOfficial: false,
-    name: "Nike ACG Polartec Wolf Tree Pullover Hoodie Anthracite - Asia",
-    price: "150,000",
-    saleCount: 250,
-    saleType: true,
-    src: "/images/newLowestAsks/img3.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 4,
-    brand: "Nike",
-    brandOfficial: false,
-    name: "Nike Air Kukini Worn Blue Obsidian",
-    price: "90,000",
-    saleCount: 422,
-    saleType: true,
-    src: "/images/newLowestAsks/img4.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 5,
-    brand: "Nike",
-    brandOfficial: false,
-    name: "Nike Kobe 4 Protro Black University Gold",
-    price: "249,000",
-    saleCount: 426,
-    saleType: true,
-    src: "/images/newLowestAsks/img5.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 6,
-    brand: "Nike",
-    brandOfficial: false,
-    name: "(W) Nike Daybreak Summit White Black",
-    price: "44,000",
-    saleCount: 6697,
-    saleType: true,
-    src: "/images/newLowestAsks/img6.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 7,
-    brand: "Supreme",
-    brandOfficial: false,
-    name: "Supreme Sleeve Arc Hooded Sweatshirt Heather Grey - 23FW",
-    price: "339,000",
-    saleCount: 79,
-    saleType: true,
-    src: "/images/newLowestAsks/img7.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 8,
-    brand: "Arc'teryx",
-    brandOfficial: false,
-    name: "Arc'teryx Heliad 10 Backpack Forage",
-    price: "300,000",
-    saleCount: 516,
-    saleType: true,
-    src: "/images/newLowestAsks/img8.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 9,
-    brand: "Asics",
-    brandOfficial: false,
-    name: "Asics x Kith Gel-Kayano 14 Cream Scarab",
-    price: "379,000",
-    saleCount: 168,
-    saleType: true,
-    src: "/images/newLowestAsks/img9.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-]
-
-export default newLowestAsksData;
