--- src/northFaceData.js
@@ -1,184 +1,0 @@
-const northFaceData = [
-  {
-    id: 0,
-    brand: "The North Face",
-    brandOfficial: false,
-    name: "The North Face White Label Kulane Down Bomber Black",
-    price: "329,000",
-    saleCount: 11,
-    saleType: true,
-    src: "/images/northFace/img0.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 1,
-    brand: "The North Face",
-    brandOfficial: false,
-    name: "(W) The North Face Greenland Windstopper Jacket Burgandy",
-    price: "240,000",
-    saleCount: 0,
-    saleType: true,
-    src: "/images/northFace/img1.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 2,
-    brand: "The North Face",
-    brandOfficial: false,
-    name: "The North Face Play Green Rvs Fleece Anorak Ivory",
-    price: "90,000",
-    saleCount: 1,
-    saleType: true,
-    src: "/images/northFace/img2.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 3,
-    brand: "The North Face",
-    brandOfficial: false,
-    name: "(W) The North Face Furry Fleece Jacket Black",
-    price: "115,000",
-    saleCount: 0,
-    saleType: true,
-    src: "/images/northFace/img3.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 4,
-    brand: "The North Face",
-    brandOfficial: false,
-    name: "The North Face Denali Wind Jacket Khaki",
-    price: "175,000",
-    saleCount: 1,
-    saleType: true,
-    src: "/images/northFace/img4.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 5,
-    brand: "The North Face",
-    brandOfficial: false,
-    name: "The North Face All-Round Bomber White Sand",
-    price: "155,000",
-    saleCount: 0,
-    saleType: true,
-    src: "/images/northFace/img5.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 6,
-    brand: "The North Face",
-    brandOfficial: false,
-    name: "The North Face Flyhigh Jacket Ink",
-    price: "90,000",
-    saleCount: 0,
-    saleType: true,
-    src: "/images/northFace/img6.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 7,
-    brand: "The North Face",
-    brandOfficial: false,
-    name: "The North Face White Label Neilton Jacket White Sand",
-    price: "169,000",
-    saleCount: 89,
-    saleType: true,
-    src: "/images/northFace/img7.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 8,
-    brand: "The North Face",
-    brandOfficial: false,
-    name: "The North Face Ice Run Jacket Black",
-    price: "168,000",
-    saleCount: 196,
-    saleType: true,
-    src: "/images/northFace/img8.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 9,
-    brand: "The North Face",
-    brandOfficial: false,
-    name: "(W) The North Face White Label Neilton Jacket Brown",
-    price: "144,000",
-    saleCount: 35,
-    saleType: true,
-    src: "/images/northFace/img9.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 10,
-    brand: "The North Face",
-    brandOfficial: false,
-    name: "The North Face Wax Heat Coach Jacket Dark Olive",
-    price: "219,000",
-    saleCount: 3,
-    saleType: true,
-    src: "/images/northFace/img10.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 11,
-    brand: "The North Face",
-    brandOfficial: false,
-    name: "The North Face Novelty GTX Mountain Jacket S Real Black",
-    price: "206,000",
-    saleCount: 3,
-    saleType: true,
-    src: "/images/northFace/img11.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 12,
-    brand: "The North Face",
-    brandOfficial: false,
-    name: "The North Face White Label Comfy Alpha Fleece Zip-Up Light Gray",
-    price: "119,000",
-    saleCount: 85,
-    saleType: true,
-    src: "/images/northFace/img12.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 13,
-    brand: "The North Face",
-    brandOfficial: false,
-    name: "The North Face White Label Rimo Reversible Fleece Hoodie Beige",
-    price: "300,000",
-    saleCount: 33,
-    saleType: true,
-    src: "/images/northFace/img13.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 14,
-    brand: "The North Face",
-    brandOfficial: false,
-    name: "(W) The North Face White Label Neilton Jacket Black",
-    price: "143,000",
-    saleCount: 13,
-    saleType: true,
-    src: "/images/northFace/img14.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-]
-
-export default northFaceData;
