--- src/nylonBagsData.js
@@ -1,184 +1,0 @@
-const nylonBagsData = [
-  {
-    id: 0,
-    brand: "Original Sports",
-    brandOfficial: true,
-    name: "Original Sports Nylon String Bag White",
-    price: "80,100",
-    saleCount: 11,
-    saleType: false,
-    src: "/images/nylonBags/img0.jpeg",
-    imgType: "jpeg",
-    bookmarkState: false
-  },
-  {
-    id: 1,
-    brand: "Balenciaga",
-    brandOfficial: false,
-    name: "Balenciaga Recycled Nylon Army Medium Tote Bag Black",
-    price: "1,130,000",
-    saleCount: 4,
-    saleType: true,
-    src: "/images/nylonBags/img1.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 2,
-    brand: "Supreme",
-    brandOfficial: false,
-    name: "Supreme Shoulder Bag Red - 22FW",
-    price: "186,000",
-    saleCount: 15,
-    saleType: true,
-    src: "/images/nylonBags/img2.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 3,
-    brand: "Icyto",
-    brandOfficial: true,
-    name: "Icyto Logo String Backpack White",
-    price: "99,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/nylonBags/img3.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 4,
-    brand: "Prada",
-    brandOfficial: false,
-    name: "Prada Re-Nylon and Saffiano Leather Duffle Bag Black",
-    price: "2,199,000",
-    saleCount: 5,
-    saleType: true,
-    src: "/images/nylonBags/img4.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 5,
-    brand: "Notabag",
-    brandOfficial: false,
-    name: "Notabag Original Navy Blue",
-    price: "25,000",
-    saleCount: 0,
-    saleType: true,
-    src: "/images/nylonBags/img5.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 6,
-    brand: "Adidas",
-    brandOfficial: false,
-    name: "Adidas Satin Mini Bucket Backpack Black",
-    price: "99,000",
-    saleCount: 77,
-    saleType: true,
-    src: "/images/nylonBags/img6.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 7,
-    brand: "Bottega Veneta",
-    brandOfficial: false,
-    name: "Bottega Veneta Padded Tech Cassette Frost",
-    price: "2,100,000",
-    saleCount: 15,
-    saleType: true,
-    src: "/images/nylonBags/img7.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 8,
-    brand: "COS",
-    brandOfficial: false,
-    name: "COS Mini Crossbody Pouch Nylon Black",
-    price: "51,000",
-    saleCount: 250,
-    saleType: true,
-    src: "/images/nylonBags/img8.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 9,
-    brand: "Vivienne Westwood",
-    brandOfficial: false,
-    name: "Vivienne Westwood Re-Nylon Drawstring Pouch Set Shoulder Bag Black",
-    price: "364,000",
-    saleCount: 27,
-    saleType: true,
-    src: "/images/nylonBags/img9.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 10,
-    brand: "Adidas",
-    brandOfficial: false,
-    name: "Adidas Adicolor Classic Festival Bag Blue Bird",
-    price: "25,000",
-    saleCount: 0,
-    saleType: true,
-    src: "/images/nylonBags/img10.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 11,
-    brand: "Stone Island",
-    brandOfficial: false,
-    name: "Stone Island 90870 Mussola Gommata Canvas Small Pouch Bag Lead - 23FW",
-    price: "298,000",
-    saleCount: 3,
-    saleType: true,
-    src: "/images/nylonBags/img11.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 12,
-    brand: "Bally",
-    brandOfficial: false,
-    name: "Bally Ferey Nylon Backpack Black",
-    price: "589,000",
-    saleCount: 2,
-    saleType: true,
-    src: "/images/nylonBags/img12.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 13,
-    brand: "Gucci",
-    brandOfficial: false,
-    name: "Gucci Micro GG Ssima Canvas Messenger Bag Black",
-    price: "709,000",
-    saleCount: 6,
-    saleType: true,
-    src: "/images/nylonBags/img13.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 14,
-    brand: "Porter",
-    brandOfficial: false,
-    name: "Porter x Comme des Garcons Backpack L Black",
-    price: "429,000",
-    saleCount: 68,
-    saleType: true,
-    src: "/images/nylonBags/img14.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-]
-
-export default nylonBagsData;
