--- src/oyData.js
@@ -1,184 +1,0 @@
-const oyData= [
-  {
-    id: 0,
-    brand: "OY",
-    brandOfficial: true,
-    name: "OY S.R Biker Jacket Black",
-    price: "179,100",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/oy/img0.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 1,
-    brand: "OY",
-    brandOfficial: true,
-    name: "OY W.S.L Track Jacket Black",
-    price: "152,100",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/oy/img1.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 2,
-    brand: "OY",
-    brandOfficial: true,
-    name: "OY Nylon Wind Breaker Jacket Black",
-    price: "134,100",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/oy/img2.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 3,
-    brand: "OY",
-    brandOfficial: true,
-    name: "OY Vintage Vegan Leather Hoodie Trucker Jacket Black",
-    price: "197,100",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/oy/img3.png",
-    imgType: "png",
-    bookmarkState: false
-  },
-  {
-    id: 4,
-    brand: "OY",
-    brandOfficial: true,
-    name: "OY Sporty Opyd Hoodie L.S T Black",
-    price: "107,100",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/oy/img4.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 5,
-    brand: "OY",
-    brandOfficial: true,
-    name: "OY Layered Hoodie Jacket Black",
-    price: "215,100",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/oy/img5.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 6,
-    brand: "OY",
-    brandOfficial: true,
-    name: "OY Heart Stitch Destroyed Knit Black",
-    price: "107,100",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/oy/img6.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 7,
-    brand: "OY",
-    brandOfficial: true,
-    name: "OY Layered Check Shirts Hoodie Black",
-    price: "134,100",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/oy/img7.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 8,
-    brand: "OY",
-    brandOfficial: true,
-    name: "OY Fourteen Cargo Pockets Washed Jeans Black",
-    price: "178,200",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/oy/img8.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 9,
-    brand: "OY",
-    brandOfficial: true,
-    name: "OY W.S.L Tape Track Pants Black",
-    price: "134,100",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/oy/img9.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 10,
-    brand: "OY",
-    brandOfficial: true,
-    name: "OY C.P Odolly Cap Black",
-    price: "44,100",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/oy/img10.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 11,
-    brand: "OY",
-    brandOfficial: true,
-    name: "OY Metal Logo Beanie Black",
-    price: "32,400",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/oy/img11.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 12,
-    brand: "OY",
-    brandOfficial: true,
-    name: "OY S.L Beanie Paul Blue",
-    price: "26,100",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/oy/img12.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 13,
-    brand: "OY",
-    brandOfficial: true,
-    name: "OY Biker Cube Keychain",
-    price: "53,100",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/oy/img13.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 14,
-    brand: "OY",
-    brandOfficial: true,
-    name: "OY Logo Keyring",
-    price: "26,100",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/oy/img14.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-]
-
-export default oyData;
