--- src/topBrandData.js
@@ -1,93 +1,0 @@
-const topBrandDetailData = [
-  {
-    id: 0,
-    type: "jpg",
-    src: "/images/brands/Nike.jpg",
-    content: "나이키"
-  },
-  {
-    id: 1,
-    type: "jpg",
-    src: "/images/brands/Stussy.jpg",
-    content: "스투시"
-  },
-  {
-    id: 2,
-    type: "jpg",
-    src: "/images/brands/Supreme.jpg",
-    content: "슈프림"
-  },
-  {
-    id: 3,
-    type: "jpg",
-    src: "/images/brands/Jordan.jpg",
-    content: "조던"
-  },
-  {
-    id: 4,
-    type: "jpg",
-    src: "/images/brands/NewBalance.jpg",
-    content: "뉴발란스"
-  },
-  {
-    id: 5,
-    type: "jpg",
-    src: "/images/brands/ArcTeryx.jpg",
-    content: "아크테릭스"
-  },
-  {
-    id: 6,
-    type: "jpg",
-    src: "/images/brands/Asics.jpg",
-    content: "아식스"
-  },
-  {
-    id: 7,
-    type: "jpg",
-    src: "/images/brands/CarharttWIP.jpg",
-    content: "칼하트 WIP"
-  },
-  {
-    id: 8,
-    type: "jpg",
-    src: "/images/brands/HumanMade.jpg",
-    content: "휴먼 메이드"
-  },
-  {
-    id: 9,
-    type: "jpg",
-    src: "/images/brands/Adidas.jpg",
-    content: "아디다스"
-  },
-  {
-    id: 10,
-    type: "jpg",
-    src: "/images/brands/IABStudio.jpg",
-    content: "아이앱 스튜디오"
-  },
-  {
-    id: 11,
-    type: "jpg",
-    src: "/images/brands/AderError.jpg",
-    content: "아더에러"
-  },
-  {
-    id: 12,
-    type: "jpg",
-    src: "/images/brands/Palace.jpg",
-    content: "팔라스"
-  },
-  {
-    id: 13,
-    type: "jpg",
-    src: "/images/brands/StoneIsland.jpg",
-    content: "스톤 아일랜드"
-  },
-  {
-    id: 14,
-    type: "jpg",
-    src: "/images/brands/PoloRalphLauren.jpg",
-    content: "폴로 랄프 로렌"
-  }
-]
-export default topBrandDetailData;
