--- src/upcomingReleaseData.js
@@ -1,164 +1,0 @@
-const upcomingReleaseData= [
-  {
-    id: 0,
-    brand: "Adidas",
-    name: "Adidas Gazelle Indoor Cloud White Collegiate Green",
-    src: "/images/upcomingRelease/img0.png",
-    imgType: "png",
-    upcomingDate: 1
-  },
-  {
-    id: 1,
-    brand: "New Balance",
-    name: "New Balance x JJJJound 2002R Gore-Tex Pine",
-    src: "/images/upcomingRelease/img1.png",
-    imgType: "png",
-    upcomingDate: 1
-  },
-  {
-    id: 2,
-    brand: "New Balance",
-    name: "New Balance x JJJJound 2002R Gore-Tex Charcoal",
-    src: "/images/upcomingRelease/img2.png",
-    imgType: "png",
-    upcomingDate: 1
-  },
-  {
-    id: 3,
-    brand: "Nike",
-    name: "Nike Air Zoom Alphafly Next% 3 Volt",
-    src: "/images/upcomingRelease/img3.png",
-    imgType: "png",
-    upcomingDate: 1
-  },
-  {
-    id: 4,
-    brand: "New Balance",
-    name: "New Balance 998 Made in USA Black Sliver",
-    src: "/images/upcomingRelease/img4.png",
-    imgType: "png",
-    upcomingDate: 1
-  },
-  {
-    id: 5,
-    brand: "Asics",
-    name: "Asics Gel-Kayano 14 Black Pure Silver",
-    src: "/images/upcomingRelease/img5.png",
-    imgType: "png",
-    upcomingDate: 1
-  },
-  {
-    id: 6,
-    brand: "Adidas",
-    name: "Adidas Adistar Cushion Cloud White Silver Metallic",
-    src: "/images/upcomingRelease/img6.png",
-    imgType: "png",
-    upcomingDate: 1
-  },
-  {
-    id: 7,
-    brand: "Clarks",
-    name: "Clarks x Martine Rose Coming Up Roses Torhill HI Green Snake",
-    src: "/images/upcomingRelease/img7.png",
-    imgType: "png",
-    upcomingDate: 2
-  },
-  {
-    id: 8,
-    brand: "Clarks",
-    name: "Clarks x Martine Rose Coming Up Roses Torhill HI Brown Snake",
-    src: "/images/upcomingRelease/img8.png",
-    imgType: "png",
-    upcomingDate: 2
-  },
-  {
-    id: 9,
-    brand: "New Balance",
-    name: "New Balance WRPD Runner Grey",
-    src: "/images/upcomingRelease/img9.png",
-    imgType: "png",
-    upcomingDate: 2
-  },
-  {
-    id: 10,
-    brand: "Nike",
-    name: "(W) Nike Air Footscape Woven Earth Light British Tan",
-    src: "/images/upcomingRelease/img10.png",
-    imgType: "png",
-    upcomingDate: 2
-  },
-  {
-    id: 11,
-    brand: "Jordan",
-    name: "Jordan 1 High 85 Team Red",
-    src: "/images/upcomingRelease/img11.png",
-    imgType: "png",
-    upcomingDate: 2
-  },
-  {
-    id: 12,
-    brand: "Clarks",
-    name: "Clarks x Martine Rose Coming Up Roses Oxford 2 Green Textile",
-    src: "/images/upcomingRelease/img12.png",
-    imgType: "png",
-    upcomingDate: 2
-  },
-  {
-    id: 13,
-    brand: "Clarks",
-    name: "Clarks x Martine Rose Coming Up Roses Oxford 2 Brown Textile",
-    src: "/images/upcomingRelease/img13.png",
-    imgType: "png",
-    upcomingDate: 2
-  },
-  {
-    id: 14,
-    brand: "New Balance",
-    name: "New Balance WRPD Runner Magnet",
-    src: "/images/upcomingRelease/img14.png",
-    imgType: "png",
-    upcomingDate: 2
-  },
-  {
-    id: 15,
-    brand: "Clarks",
-    name: "Clarks x Martine Rose Coming Up Roses Oxford 2 Black Textile",
-    src: "/images/upcomingRelease/img15.png",
-    imgType: "png",
-    upcomingDate: 2
-  },
-  {
-    id: 16,
-    brand: "Adidas",
-    name: "(C) Adidas x Bad Bunny Response CL Wonder White Clear Granite",
-    src: "/images/upcomingRelease/img16.png",
-    imgType: "png",
-    upcomingDate: 3
-  },
-  {
-    id: 17,
-    brand: "Palace",
-    name: "Palace Horsepower Trucker Black - 24SS",
-    src: "/images/upcomingRelease/img17.png",
-    imgType: "png",
-    upcomingDate: 3
-  },
-  {
-    id: 18,
-    brand: "Jordan",
-    name: "(PS) Jordan 4 Retro Bred Reimagined",
-    src: "/images/upcomingRelease/img18.png",
-    imgType: "png",
-    upcomingDate: 3
-  },
-  {
-    id: 19,
-    brand: "Palace",
-    name: "Palace Teddy Ears Earflapper Black - 24SS",
-    src: "/images/upcomingRelease/img19.png",
-    imgType: "png",
-    upcomingDate: 3
-  },
-]
-
-export default upcomingReleaseData;
