--- src/weeklyPickData.js
@@ -1,184 +1,0 @@
-const weeklyPickData= [
-  {
-    id: 0,
-    brand: "Ottolinger",
-    brandOfficial: true,
-    name: "Ottolinger Women Woven Denim Drape Pants Blue",
-    price: "382,500",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/weeklyPick/img0.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 1,
-    brand: "Woolrich",
-    brandOfficial: true,
-    name: "Woolrich Gtx 2L Fishing Gear Jacket Beige",
-    price: "393,500",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/weeklyPick/img1.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 2,
-    brand: "Ottolinger",
-    brandOfficial: true,
-    name: "Ottolinger Women Signature Split Blazer Black Pinstripe",
-    price: "1,300,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/weeklyPick/img2.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 3,
-    brand: "The Garment",
-    brandOfficial: true,
-    name: "The Garment Women Valetta Shirt Blouse Vanilla",
-    price: "460,700",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/weeklyPick/img3.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 4,
-    brand: "Woolrich",
-    brandOfficial: true,
-    name: "Woolrich Gtx 3L Fishing Gear Jacket Olive",
-    price: "458,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/weeklyPick/img4.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 5,
-    brand: "Werkstatt:Munchen",
-    brandOfficial: true,
-    name: "Werkstatt:Munchen Necklace Two Chains Sterling Silver",
-    price: "545,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/weeklyPick/img5.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 6,
-    brand: "Paula Canovas Del Vas",
-    brandOfficial: true,
-    name: "Paula Canovas Del Vas Women Diablo Heel Sandals Black",
-    price: "565,200",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/weeklyPick/img6.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 7,
-    brand: "Perverze",
-    brandOfficial: true,
-    name: "Perverze Wool String Top and Scarf Red",
-    price: "204,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/weeklyPick/img7.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 8,
-    brand: "Ottolinger",
-    brandOfficial: true,
-    name: "Ottolinger Women Woven Double Waistband Suit Trousers Blue",
-    price: "550,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/weeklyPick/img8.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 9,
-    brand: "AMOMENTO",
-    brandOfficial: true,
-    name: "Amomento Mens Boucle Raglan Coat Brown",
-    price: "538,300",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/weeklyPick/img9.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 10,
-    brand: "Hype Clothing",
-    brandOfficial: true,
-    name: "Nike x Supreme Air Zoom Courtposite SP White University Red",
-    price: "328,500",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/weeklyPick/img10.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 11,
-    brand: "Werkstatt:Munchen",
-    brandOfficial: true,
-    name: "Werkstatt:Munchen 4 Ring Combination Cross Sterling Silver",
-    price: "695,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/weeklyPick/img11.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 12,
-    brand: "Perverze",
-    brandOfficial: true,
-    name: "Perverze Narrow Tie T-Shiet White",
-    price: "150,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/weeklyPick/img12.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 13,
-    brand: "Yunse",
-    brandOfficial: true,
-    name: "Yunse Women Unbalance Pleats Skirt Black",
-    price: "242,200",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/weeklyPick/img13.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-  {
-    id: 14,
-    brand: "Hoka",
-    brandOfficial: true,
-    name: "Hoka Kaha 2 Low GTX Oxford Tan Dune",
-    price: "279,000",
-    saleCount: 0,
-    saleType: false,
-    src: "/images/weeklyPick/img14.jpg",
-    imgType: "jpg",
-    bookmarkState: false
-  },
-]
-
-export default weeklyPickData;
